forge
Module backend

Foundry’s main executor backend abstraction and implementation.

Structs

Backend

Provides the underlying revm::Database implementation.

BackendInner

Container type for various Backend related data

BackendStateSnapshot

Represents a state snapshot taken during evm execution

BlockchainDb

A shareable Block database

BlockchainDbMeta

relevant identifying markers in the context of BlockchainDb

CowBackend

A wrapper around Backend that ensures only revm::DatabaseRef functions are called.

EmptyDBWrapper

An empty database that always returns default values when queried.

Fork

Represents a fork

MemDb

In-memory Database for Anvil.

SharedBackend

A cloneable backend type that shares access to the backend data with all its clones.

StateSnapshot

A minimal abstraction of a state at a certain point in time

Enums

BackendDatabaseSnapshot

Variants of a revm::Database

BackendError

Errors that can happen when working with revm::Database

DatabaseError

Errors that can happen when working with revm::Database

RevertDiagnostic

Represents possible diagnostic cases on revert

RevertStateSnapshotAction

What to do when reverting a state snapshot.

Constants

GLOBAL_FAIL_SLOT

bytes32("failed"), as a storage slot key into CHEATCODE_ADDRESS.

Traits

DatabaseExt

An extension trait that allows us to easily extend the revm::Inspector capabilities

Functions

update_state

Helper method which updates data in the state with the data from the database. Does not change state for persistent accounts (for roll fork to transaction and transact).

Type Aliases

BackendResult

DatabaseResult

Result alias with DatabaseError as error

FoundryEvmInMemoryDB

Type alias for an in-memory database.

LocalForkId

Represents a numeric ForkId valid only for the existence of the Backend.