forge

Module cmd

forge subcommands.

All subcommands should respect the foundry_config::Config. If a subcommand accepts values that are supported by the Config, then the subcommand should implement figment::Provider which allows the subcommand to override the config’s defaults, see [foundry_config::Config].

Modules

bind

bind_json

build

cache

clone

compiler

config

coverage

create

doc

eip712

flatten

fmt

geiger

generate

init

inspect

install

lint

remappings

remove

selectors

snapshot

soldeer

test

tree

update

watch