forge

Crate core

foundry-evm-core

Core EVM abstractions.

Modules

abi

backend

Foundry’s main executor backend abstraction and implementation.

buffer

bytecode

constants

decode

Various utilities to decode test results.

either_evm

env

evm

fork

ic

opts

precompiles

state_snapshot

Support for snapshotting different states

utils

Structs

Env

Helper container type for EvmEnv and TxEnv.

EnvMut

Helper struct with mutable references to the block and cfg environments.

EvmEnv

Container type that holds both the configuration and block environment for EVM execution.

Traits

AsEnvMut

ContextExt

InspectorExt

An extension trait that allows us to add additional hooks to Inspector for later use in handlers.

Type Aliases

Breakpoints

Map keyed by breakpoints char to their location (contract address, pc)

Ecx