Module forge::coverage

Coverage reports.

Modules

analysis

anchors

Structs

BytecodeReporter

ContractId

A unique identifier for a contract

CoverageCollector

CoverageItem

CoverageReport

A coverage report.

CoverageSummary

Coverage summary for a source file.

DebugReporter

A super verbose reporter for debugging coverage while it is still unstable.

HitMap

Hit data for an address.

HitMaps

A collection of HitMaps

ItemAnchor

An item anchor describes what instruction marks a CoverageItem as covered.

LcovReporter

LineNumberCache 🔒

Cache line number offsets for source files

SourceLocation

SummaryReporter

A simple summary reporter that prints the coverage results in a table.

Enums

CoverageItemKind

Traits

CoverageReporter

A coverage reporter.

Functions

format_cell 🔒