forge

Module executors

EVM executor abstractions, which can execute calls.

Used for running tests, scripts, and interacting with the inner backend which holds the state.

Modules

ITest

Module containing a contract’s types and functions.

fuzz

invariant

strategy

Structs

CallResult

The result of a call.

DeployResult

The result of a deployment.

EarlyExit

Helper struct to enable early exit behavior: when one test fails or run is interrupted, all other tests stop early.

EvmExecutorStrategyRunner

Implements ExecutorStrategyRunner for EVM.

ExecutionErr

Represents the context after an execution error occurred.

Executor

EVM executor.

ExecutorBuilder

The builder that allows to configure an evm Executor which a stack of optional revm::Inspectors, such as Cheatcodes.

ExecutorStrategy

Defines the strategy for an Executor.

FuzzTestTimer

Timer for a fuzz test.

FuzzedExecutor

Wrapper around an Executor which provides fuzzing support using [proptest].

InvariantExecutor

Wrapper around any Executor implementer which provides fuzzing support using [proptest].

RawCallResult

The result of a raw call.

TracingExecutor

A default executor with tracing enabled

Enums

EvmError

Traits

ExecutorStrategyContext

Context for ExecutorStrategy.

ExecutorStrategyExt

Extended trait for Polkadot.

ExecutorStrategyRunner

Stateless strategy runner for ExecutorStrategy.