forge::executors

Module invariant

Modules

IInvariantTest

Module containing a contract’s types and functions.

Structs

InvariantExecutor

Wrapper around any Executor implementer which provides fuzzing support using [proptest].

InvariantFailures

Stores information about failures and reverts of the invariant tests.

InvariantFuzzTestResult

The outcome of an invariant fuzz test

InvariantMetrics

Contains invariant metrics for a single fuzzed selector.

Enums

InvariantFuzzError

Functions

check_sequence

Checks if the given call sequence breaks the invariant.

replay_error

Replays the error case, shrinks the failing sequence and collects all necessary traces.

replay_run

Replays a call sequence for collecting logs and traces. Returns counterexample to be used when the call sequence is a failed scenario.