forge::fuzz::strategies

Trait BoundMutator

pub trait BoundMutator:
    Sized
    + Copy
    + Debug {
    // Required method
    fn bound(
        self,
        min: Self,
        max: Self,
        test_runner: &mut TestRunner,
    ) -> Option<Self>;
}

Mutator that bounds the current value of an uint or int in the given range. The mutated value is always different from the current value.

Required Methods

fn bound( self, min: Self, max: Self, test_runner: &mut TestRunner, ) -> Option<Self>

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety".