forge::inspectors::cheatcodes::spec::Vm

Enum VmCalls

pub enum VmCalls {

    addr(addrCall),
    dumpState(dumpStateCall),
    getNonce_0(getNonce_0Call),
    getNonce_1(getNonce_1Call),
    load(loadCall),
    loadAllocs(loadAllocsCall),
    startDebugTraceRecording(startDebugTraceRecordingCall),
    stopAndReturnDebugTraceRecording(stopAndReturnDebugTraceRecordingCall),
    cloneAccount(cloneAccountCall),
    record(recordCall),
    stopRecord(stopRecordCall),
    accesses(accessesCall),
    startStateDiffRecording(startStateDiffRecordingCall),
    stopAndReturnStateDiff(stopAndReturnStateDiffCall),
    getStateDiff(getStateDiffCall),
    getStateDiffJson(getStateDiffJsonCall),
    getStorageAccesses(getStorageAccessesCall),
    startMappingRecording(startMappingRecordingCall),
    stopMappingRecording(stopMappingRecordingCall),
    getMappingLength(getMappingLengthCall),
    getMappingSlotAt(getMappingSlotAtCall),
    getMappingKeyAndParentOf(getMappingKeyAndParentOfCall),
    chainId(chainIdCall),
    coinbase(coinbaseCall),
    difficulty(difficultyCall),
    fee(feeCall),
    prevrandao_0(prevrandao_0Call),
    prevrandao_1(prevrandao_1Call),
    blobhashes(blobhashesCall),
    getBlobhashes(getBlobhashesCall),
    roll(rollCall),
    getBlockNumber(getBlockNumberCall),
    txGasPrice(txGasPriceCall),
    warp(warpCall),
    getBlockTimestamp(getBlockTimestampCall),
    getRawBlockHeader(getRawBlockHeaderCall),
    blobBaseFee(blobBaseFeeCall),
    getBlobBaseFee(getBlobBaseFeeCall),
    setBlockhash(setBlockhashCall),
    deal(dealCall),
    etch(etchCall),
    resetNonce(resetNonceCall),
    setNonce(setNonceCall),
    setNonceUnsafe(setNonceUnsafeCall),
    store(storeCall),
    cool(coolCall),
    accessList(accessListCall),
    noAccessList(noAccessListCall),
    warmSlot(warmSlotCall),
    coolSlot(coolSlotCall),
    clearMockedCalls(clearMockedCallsCall),
    mockCall_0(mockCall_0Call),
    mockCall_1(mockCall_1Call),
    mockCall_2(mockCall_2Call),
    mockCall_3(mockCall_3Call),
    mockCalls_0(mockCalls_0Call),
    mockCalls_1(mockCalls_1Call),
    mockCallRevert_0(mockCallRevert_0Call),
    mockCallRevert_1(mockCallRevert_1Call),
    mockCallRevert_2(mockCallRevert_2Call),
    mockCallRevert_3(mockCallRevert_3Call),
    mockFunction(mockFunctionCall),
    prank_0(prank_0Call),
    startPrank_0(startPrank_0Call),
    prank_1(prank_1Call),
    startPrank_1(startPrank_1Call),
    prank_2(prank_2Call),
    startPrank_2(startPrank_2Call),
    prank_3(prank_3Call),
    startPrank_3(startPrank_3Call),
    stopPrank(stopPrankCall),
    readCallers(readCallersCall),
    snapshotValue_0(snapshotValue_0Call),
    snapshotValue_1(snapshotValue_1Call),
    snapshotGasLastCall_0(snapshotGasLastCall_0Call),
    snapshotGasLastCall_1(snapshotGasLastCall_1Call),
    startSnapshotGas_0(startSnapshotGas_0Call),
    startSnapshotGas_1(startSnapshotGas_1Call),
    stopSnapshotGas_0(stopSnapshotGas_0Call),
    stopSnapshotGas_1(stopSnapshotGas_1Call),
    stopSnapshotGas_2(stopSnapshotGas_2Call),
    snapshot(snapshotCall),
    snapshotState(snapshotStateCall),
    revertTo(revertToCall),
    revertToState(revertToStateCall),
    revertToAndDelete(revertToAndDeleteCall),
    revertToStateAndDelete(revertToStateAndDeleteCall),
    deleteSnapshot(deleteSnapshotCall),
    deleteStateSnapshot(deleteStateSnapshotCall),
    deleteSnapshots(deleteSnapshotsCall),
    deleteStateSnapshots(deleteStateSnapshotsCall),
    activeFork(activeForkCall),
    createFork_0(createFork_0Call),
    createFork_1(createFork_1Call),
    createFork_2(createFork_2Call),
    createSelectFork_0(createSelectFork_0Call),
    createSelectFork_1(createSelectFork_1Call),
    createSelectFork_2(createSelectFork_2Call),
    rollFork_0(rollFork_0Call),
    rollFork_1(rollFork_1Call),
    rollFork_2(rollFork_2Call),
    rollFork_3(rollFork_3Call),
    selectFork(selectForkCall),
    transact_0(transact_0Call),
    transact_1(transact_1Call),
    rpc_0(rpc_0Call),
    rpc_1(rpc_1Call),
    eth_getLogs(eth_getLogsCall),
    allowCheatcodes(allowCheatcodesCall),
    makePersistent_0(makePersistent_0Call),
    makePersistent_1(makePersistent_1Call),
    makePersistent_2(makePersistent_2Call),
    makePersistent_3(makePersistent_3Call),
    revokePersistent_0(revokePersistent_0Call),
    revokePersistent_1(revokePersistent_1Call),
    isPersistent(isPersistentCall),
    recordLogs(recordLogsCall),
    getRecordedLogs(getRecordedLogsCall),
    pauseGasMetering(pauseGasMeteringCall),
    resumeGasMetering(resumeGasMeteringCall),
    resetGasMetering(resetGasMeteringCall),
    lastCallGas(lastCallGasCall),
    assume(assumeCall),
    assumeNoRevert_0(assumeNoRevert_0Call),
    assumeNoRevert_1(assumeNoRevert_1Call),
    assumeNoRevert_2(assumeNoRevert_2Call),
    breakpoint_0(breakpoint_0Call),
    breakpoint_1(breakpoint_1Call),
    getFoundryVersion(getFoundryVersionCall),
    rpcUrl(rpcUrlCall),
    rpcUrls(rpcUrlsCall),
    rpcUrlStructs(rpcUrlStructsCall),
    getChain_0(getChain_0Call),
    getChain_1(getChain_1Call),
    sleep(sleepCall),
    expectCall_0(expectCall_0Call),
    expectCall_1(expectCall_1Call),
    expectCall_2(expectCall_2Call),
    expectCall_3(expectCall_3Call),
    expectCall_4(expectCall_4Call),
    expectCall_5(expectCall_5Call),
    expectCallMinGas_0(expectCallMinGas_0Call),
    expectCallMinGas_1(expectCallMinGas_1Call),
    expectEmit_0(expectEmit_0Call),
    expectEmit_1(expectEmit_1Call),
    expectEmit_2(expectEmit_2Call),
    expectEmit_3(expectEmit_3Call),
    expectEmit_4(expectEmit_4Call),
    expectEmit_5(expectEmit_5Call),
    expectEmit_6(expectEmit_6Call),
    expectEmit_7(expectEmit_7Call),
    expectEmitAnonymous_0(expectEmitAnonymous_0Call),
    expectEmitAnonymous_1(expectEmitAnonymous_1Call),
    expectEmitAnonymous_2(expectEmitAnonymous_2Call),
    expectEmitAnonymous_3(expectEmitAnonymous_3Call),
    expectCreate(expectCreateCall),
    expectCreate2(expectCreate2Call),
    expectRevert_0(expectRevert_0Call),
    expectRevert_1(expectRevert_1Call),
    expectRevert_2(expectRevert_2Call),
    expectRevert_3(expectRevert_3Call),
    expectRevert_4(expectRevert_4Call),
    expectRevert_5(expectRevert_5Call),
    expectRevert_6(expectRevert_6Call),
    expectRevert_7(expectRevert_7Call),
    expectRevert_8(expectRevert_8Call),
    expectRevert_9(expectRevert_9Call),
    expectRevert_10(expectRevert_10Call),
    expectRevert_11(expectRevert_11Call),
    expectPartialRevert_0(expectPartialRevert_0Call),
    expectPartialRevert_1(expectPartialRevert_1Call),
    _expectCheatcodeRevert_0(_expectCheatcodeRevert_0Call),
    _expectCheatcodeRevert_1(_expectCheatcodeRevert_1Call),
    _expectCheatcodeRevert_2(_expectCheatcodeRevert_2Call),
    expectSafeMemory(expectSafeMemoryCall),
    stopExpectSafeMemory(stopExpectSafeMemoryCall),
    expectSafeMemoryCall(expectSafeMemoryCallCall),
    skip_0(skip_0Call),
    skip_1(skip_1Call),
    assertTrue_0(assertTrue_0Call),
    assertTrue_1(assertTrue_1Call),
    assertFalse_0(assertFalse_0Call),
    assertFalse_1(assertFalse_1Call),
    assertEq_0(assertEq_0Call),
    assertEq_1(assertEq_1Call),
    assertEq_2(assertEq_2Call),
    assertEq_3(assertEq_3Call),
    assertEq_4(assertEq_4Call),
    assertEq_5(assertEq_5Call),
    assertEq_6(assertEq_6Call),
    assertEq_7(assertEq_7Call),
    assertEq_8(assertEq_8Call),
    assertEq_9(assertEq_9Call),
    assertEq_10(assertEq_10Call),
    assertEq_11(assertEq_11Call),
    assertEq_12(assertEq_12Call),
    assertEq_13(assertEq_13Call),
    assertEq_14(assertEq_14Call),
    assertEq_15(assertEq_15Call),
    assertEq_16(assertEq_16Call),
    assertEq_17(assertEq_17Call),
    assertEq_18(assertEq_18Call),
    assertEq_19(assertEq_19Call),
    assertEq_20(assertEq_20Call),
    assertEq_21(assertEq_21Call),
    assertEq_22(assertEq_22Call),
    assertEq_23(assertEq_23Call),
    assertEq_24(assertEq_24Call),
    assertEq_25(assertEq_25Call),
    assertEq_26(assertEq_26Call),
    assertEq_27(assertEq_27Call),
    assertEqDecimal_0(assertEqDecimal_0Call),
    assertEqDecimal_1(assertEqDecimal_1Call),
    assertEqDecimal_2(assertEqDecimal_2Call),
    assertEqDecimal_3(assertEqDecimal_3Call),
    assertNotEq_0(assertNotEq_0Call),
    assertNotEq_1(assertNotEq_1Call),
    assertNotEq_2(assertNotEq_2Call),
    assertNotEq_3(assertNotEq_3Call),
    assertNotEq_4(assertNotEq_4Call),
    assertNotEq_5(assertNotEq_5Call),
    assertNotEq_6(assertNotEq_6Call),
    assertNotEq_7(assertNotEq_7Call),
    assertNotEq_8(assertNotEq_8Call),
    assertNotEq_9(assertNotEq_9Call),
    assertNotEq_10(assertNotEq_10Call),
    assertNotEq_11(assertNotEq_11Call),
    assertNotEq_12(assertNotEq_12Call),
    assertNotEq_13(assertNotEq_13Call),
    assertNotEq_14(assertNotEq_14Call),
    assertNotEq_15(assertNotEq_15Call),
    assertNotEq_16(assertNotEq_16Call),
    assertNotEq_17(assertNotEq_17Call),
    assertNotEq_18(assertNotEq_18Call),
    assertNotEq_19(assertNotEq_19Call),
    assertNotEq_20(assertNotEq_20Call),
    assertNotEq_21(assertNotEq_21Call),
    assertNotEq_22(assertNotEq_22Call),
    assertNotEq_23(assertNotEq_23Call),
    assertNotEq_24(assertNotEq_24Call),
    assertNotEq_25(assertNotEq_25Call),
    assertNotEq_26(assertNotEq_26Call),
    assertNotEq_27(assertNotEq_27Call),
    assertNotEqDecimal_0(assertNotEqDecimal_0Call),
    assertNotEqDecimal_1(assertNotEqDecimal_1Call),
    assertNotEqDecimal_2(assertNotEqDecimal_2Call),
    assertNotEqDecimal_3(assertNotEqDecimal_3Call),
    assertGt_0(assertGt_0Call),
    assertGt_1(assertGt_1Call),
    assertGt_2(assertGt_2Call),
    assertGt_3(assertGt_3Call),
    assertGtDecimal_0(assertGtDecimal_0Call),
    assertGtDecimal_1(assertGtDecimal_1Call),
    assertGtDecimal_2(assertGtDecimal_2Call),
    assertGtDecimal_3(assertGtDecimal_3Call),
    assertGe_0(assertGe_0Call),
    assertGe_1(assertGe_1Call),
    assertGe_2(assertGe_2Call),
    assertGe_3(assertGe_3Call),
    assertGeDecimal_0(assertGeDecimal_0Call),
    assertGeDecimal_1(assertGeDecimal_1Call),
    assertGeDecimal_2(assertGeDecimal_2Call),
    assertGeDecimal_3(assertGeDecimal_3Call),
    assertLt_0(assertLt_0Call),
    assertLt_1(assertLt_1Call),
    assertLt_2(assertLt_2Call),
    assertLt_3(assertLt_3Call),
    assertLtDecimal_0(assertLtDecimal_0Call),
    assertLtDecimal_1(assertLtDecimal_1Call),
    assertLtDecimal_2(assertLtDecimal_2Call),
    assertLtDecimal_3(assertLtDecimal_3Call),
    assertLe_0(assertLe_0Call),
    assertLe_1(assertLe_1Call),
    assertLe_2(assertLe_2Call),
    assertLe_3(assertLe_3Call),
    assertLeDecimal_0(assertLeDecimal_0Call),
    assertLeDecimal_1(assertLeDecimal_1Call),
    assertLeDecimal_2(assertLeDecimal_2Call),
    assertLeDecimal_3(assertLeDecimal_3Call),
    assertApproxEqAbs_0(assertApproxEqAbs_0Call),
    assertApproxEqAbs_1(assertApproxEqAbs_1Call),
    assertApproxEqAbs_2(assertApproxEqAbs_2Call),
    assertApproxEqAbs_3(assertApproxEqAbs_3Call),
    assertApproxEqAbsDecimal_0(assertApproxEqAbsDecimal_0Call),
    assertApproxEqAbsDecimal_1(assertApproxEqAbsDecimal_1Call),
    assertApproxEqAbsDecimal_2(assertApproxEqAbsDecimal_2Call),
    assertApproxEqAbsDecimal_3(assertApproxEqAbsDecimal_3Call),
    assertApproxEqRel_0(assertApproxEqRel_0Call),
    assertApproxEqRel_1(assertApproxEqRel_1Call),
    assertApproxEqRel_2(assertApproxEqRel_2Call),
    assertApproxEqRel_3(assertApproxEqRel_3Call),
    assertApproxEqRelDecimal_0(assertApproxEqRelDecimal_0Call),
    assertApproxEqRelDecimal_1(assertApproxEqRelDecimal_1Call),
    assertApproxEqRelDecimal_2(assertApproxEqRelDecimal_2Call),
    assertApproxEqRelDecimal_3(assertApproxEqRelDecimal_3Call),
    foundryVersionAtLeast(foundryVersionAtLeastCall),
    foundryVersionCmp(foundryVersionCmpCall),
    exists(existsCall),
    fsMetadata(fsMetadataCall),
    isDir(isDirCall),
    isFile(isFileCall),
    projectRoot(projectRootCall),
    unixTime(unixTimeCall),
    closeFile(closeFileCall),
    copyFile(copyFileCall),
    createDir(createDirCall),
    readDir_0(readDir_0Call),
    readDir_1(readDir_1Call),
    readDir_2(readDir_2Call),
    readFile(readFileCall),
    readFileBinary(readFileBinaryCall),
    readLine(readLineCall),
    readLink(readLinkCall),
    removeDir(removeDirCall),
    removeFile(removeFileCall),
    writeFile(writeFileCall),
    writeFileBinary(writeFileBinaryCall),
    writeLine(writeLineCall),
    getArtifactPathByCode(getArtifactPathByCodeCall),
    getArtifactPathByDeployedCode(getArtifactPathByDeployedCodeCall),
    getCode(getCodeCall),
    deployCode_0(deployCode_0Call),
    deployCode_1(deployCode_1Call),
    deployCode_2(deployCode_2Call),
    deployCode_3(deployCode_3Call),
    deployCode_4(deployCode_4Call),
    deployCode_5(deployCode_5Call),
    deployCode_6(deployCode_6Call),
    deployCode_7(deployCode_7Call),
    getDeployedCode(getDeployedCodeCall),
    getBroadcast(getBroadcastCall),
    getBroadcasts_0(getBroadcasts_0Call),
    getBroadcasts_1(getBroadcasts_1Call),
    getDeployment_0(getDeployment_0Call),
    getDeployment_1(getDeployment_1Call),
    getDeployments(getDeploymentsCall),
    ffi(ffiCall),
    tryFfi(tryFfiCall),
    prompt(promptCall),
    promptSecret(promptSecretCall),
    promptSecretUint(promptSecretUintCall),
    promptAddress(promptAddressCall),
    promptUint(promptUintCall),
    setEnv(setEnvCall),
    envExists(envExistsCall),
    envBool_0(envBool_0Call),
    envUint_0(envUint_0Call),
    envInt_0(envInt_0Call),
    envAddress_0(envAddress_0Call),
    envBytes32_0(envBytes32_0Call),
    envString_0(envString_0Call),
    envBytes_0(envBytes_0Call),
    envBool_1(envBool_1Call),
    envUint_1(envUint_1Call),
    envInt_1(envInt_1Call),
    envAddress_1(envAddress_1Call),
    envBytes32_1(envBytes32_1Call),
    envString_1(envString_1Call),
    envBytes_1(envBytes_1Call),
    envOr_0(envOr_0Call),
    envOr_1(envOr_1Call),
    envOr_2(envOr_2Call),
    envOr_3(envOr_3Call),
    envOr_4(envOr_4Call),
    envOr_5(envOr_5Call),
    envOr_6(envOr_6Call),
    envOr_7(envOr_7Call),
    envOr_8(envOr_8Call),
    envOr_9(envOr_9Call),
    envOr_10(envOr_10Call),
    envOr_11(envOr_11Call),
    envOr_12(envOr_12Call),
    envOr_13(envOr_13Call),
    isContext(isContextCall),
    forkChains(forkChainsCall),
    forkChainIds(forkChainIdsCall),
    forkChain(forkChainCall),
    forkChainId(forkChainIdCall),
    forkRpcUrl(forkRpcUrlCall),
    forkChainRpcUrl(forkChainRpcUrlCall),
    forkBool(forkBoolCall),
    forkChainBool(forkChainBoolCall),
    forkInt(forkIntCall),
    forkChainInt(forkChainIntCall),
    forkUint(forkUintCall),
    forkChainUint(forkChainUintCall),
    forkAddress(forkAddressCall),
    forkChainAddress(forkChainAddressCall),
    forkBytes32(forkBytes32Call),
    forkChainBytes32(forkChainBytes32Call),
    forkString(forkStringCall),
    forkChainString(forkChainStringCall),
    forkBytes(forkBytesCall),
    forkChainBytes(forkChainBytesCall),
    broadcast_0(broadcast_0Call),
    broadcast_1(broadcast_1Call),
    broadcast_2(broadcast_2Call),
    startBroadcast_0(startBroadcast_0Call),
    startBroadcast_1(startBroadcast_1Call),
    startBroadcast_2(startBroadcast_2Call),
    stopBroadcast(stopBroadcastCall),
    broadcastRawTransaction(broadcastRawTransactionCall),
    signDelegation_0(signDelegation_0Call),
    signDelegation_1(signDelegation_1Call),
    signDelegation_2(signDelegation_2Call),
    attachDelegation_0(attachDelegation_0Call),
    attachDelegation_1(attachDelegation_1Call),
    signAndAttachDelegation_0(signAndAttachDelegation_0Call),
    signAndAttachDelegation_1(signAndAttachDelegation_1Call),
    signAndAttachDelegation_2(signAndAttachDelegation_2Call),
    attachBlob(attachBlobCall),
    getWallets(getWalletsCall),
    toString_0(toString_0Call),
    toString_1(toString_1Call),
    toString_2(toString_2Call),
    toString_3(toString_3Call),
    toString_4(toString_4Call),
    toString_5(toString_5Call),
    parseBytes(parseBytesCall),
    parseAddress(parseAddressCall),
    parseUint(parseUintCall),
    parseInt(parseIntCall),
    parseBytes32(parseBytes32Call),
    parseBool(parseBoolCall),
    toLowercase(toLowercaseCall),
    toUppercase(toUppercaseCall),
    trim(trimCall),
    replace(replaceCall),
    split(splitCall),
    indexOf(indexOfCall),
    contains(containsCall),
    keyExists(keyExistsCall),
    keyExistsJson(keyExistsJsonCall),
    parseJson_0(parseJson_0Call),
    parseJson_1(parseJson_1Call),
    parseJsonUint(parseJsonUintCall),
    parseJsonUintArray(parseJsonUintArrayCall),
    parseJsonInt(parseJsonIntCall),
    parseJsonIntArray(parseJsonIntArrayCall),
    parseJsonBool(parseJsonBoolCall),
    parseJsonBoolArray(parseJsonBoolArrayCall),
    parseJsonAddress(parseJsonAddressCall),
    parseJsonAddressArray(parseJsonAddressArrayCall),
    parseJsonString(parseJsonStringCall),
    parseJsonStringArray(parseJsonStringArrayCall),
    parseJsonBytes(parseJsonBytesCall),
    parseJsonBytesArray(parseJsonBytesArrayCall),
    parseJsonBytes32(parseJsonBytes32Call),
    parseJsonBytes32Array(parseJsonBytes32ArrayCall),
    parseJsonType_0(parseJsonType_0Call),
    parseJsonType_1(parseJsonType_1Call),
    parseJsonTypeArray(parseJsonTypeArrayCall),
    parseJsonKeys(parseJsonKeysCall),
    serializeJson(serializeJsonCall),
    serializeBool_0(serializeBool_0Call),
    serializeUint_0(serializeUint_0Call),
    serializeUintToHex(serializeUintToHexCall),
    serializeInt_0(serializeInt_0Call),
    serializeAddress_0(serializeAddress_0Call),
    serializeBytes32_0(serializeBytes32_0Call),
    serializeString_0(serializeString_0Call),
    serializeBytes_0(serializeBytes_0Call),
    serializeBool_1(serializeBool_1Call),
    serializeUint_1(serializeUint_1Call),
    serializeInt_1(serializeInt_1Call),
    serializeAddress_1(serializeAddress_1Call),
    serializeBytes32_1(serializeBytes32_1Call),
    serializeString_1(serializeString_1Call),
    serializeBytes_1(serializeBytes_1Call),
    serializeJsonType_0(serializeJsonType_0Call),
    serializeJsonType_1(serializeJsonType_1Call),
    writeJson_0(writeJson_0Call),
    writeJson_1(writeJson_1Call),
    keyExistsToml(keyExistsTomlCall),
    parseToml_0(parseToml_0Call),
    parseToml_1(parseToml_1Call),
    parseTomlUint(parseTomlUintCall),
    parseTomlUintArray(parseTomlUintArrayCall),
    parseTomlInt(parseTomlIntCall),
    parseTomlIntArray(parseTomlIntArrayCall),
    parseTomlBool(parseTomlBoolCall),
    parseTomlBoolArray(parseTomlBoolArrayCall),
    parseTomlAddress(parseTomlAddressCall),
    parseTomlAddressArray(parseTomlAddressArrayCall),
    parseTomlString(parseTomlStringCall),
    parseTomlStringArray(parseTomlStringArrayCall),
    parseTomlBytes(parseTomlBytesCall),
    parseTomlBytesArray(parseTomlBytesArrayCall),
    parseTomlBytes32(parseTomlBytes32Call),
    parseTomlBytes32Array(parseTomlBytes32ArrayCall),
    parseTomlType_0(parseTomlType_0Call),
    parseTomlType_1(parseTomlType_1Call),
    parseTomlTypeArray(parseTomlTypeArrayCall),
    parseTomlKeys(parseTomlKeysCall),
    writeToml_0(writeToml_0Call),
    writeToml_1(writeToml_1Call),
    createWallet_0(createWallet_0Call),
    createWallet_1(createWallet_1Call),
    createWallet_2(createWallet_2Call),
    sign_0(sign_0Call),
    signCompact_0(signCompact_0Call),
    sign_1(sign_1Call),
    signCompact_1(signCompact_1Call),
    sign_2(sign_2Call),
    signCompact_2(signCompact_2Call),
    sign_3(sign_3Call),
    signCompact_3(signCompact_3Call),
    signP256(signP256Call),
    publicKeyP256(publicKeyP256Call),
    deriveKey_0(deriveKey_0Call),
    deriveKey_1(deriveKey_1Call),
    deriveKey_2(deriveKey_2Call),
    deriveKey_3(deriveKey_3Call),
    rememberKey(rememberKeyCall),
    rememberKeys_0(rememberKeys_0Call),
    rememberKeys_1(rememberKeys_1Call),
    label(labelCall),
    getLabel(getLabelCall),
    computeCreateAddress(computeCreateAddressCall),
    computeCreate2Address_0(computeCreate2Address_0Call),
    computeCreate2Address_1(computeCreate2Address_1Call),
    toBase64_0(toBase64_0Call),
    toBase64_1(toBase64_1Call),
    toBase64URL_0(toBase64URL_0Call),
    toBase64URL_1(toBase64URL_1Call),
    ensNamehash(ensNamehashCall),
    randomUint_0(randomUint_0Call),
    randomUint_1(randomUint_1Call),
    randomUint_2(randomUint_2Call),
    randomAddress(randomAddressCall),
    randomInt_0(randomInt_0Call),
    randomInt_1(randomInt_1Call),
    randomBool(randomBoolCall),
    randomBytes(randomBytesCall),
    randomBytes4(randomBytes4Call),
    randomBytes8(randomBytes8Call),
    pauseTracing(pauseTracingCall),
    resumeTracing(resumeTracingCall),
    copyStorage(copyStorageCall),
    setArbitraryStorage_0(setArbitraryStorage_0Call),
    setArbitraryStorage_1(setArbitraryStorage_1Call),
    sort(sortCall),
    shuffle(shuffleCall),
    setSeed(setSeedCall),
    interceptInitcode(interceptInitcodeCall),
    eip712HashType_0(eip712HashType_0Call),
    eip712HashType_1(eip712HashType_1Call),
    eip712HashStruct_0(eip712HashStruct_0Call),
    eip712HashStruct_1(eip712HashStruct_1Call),
    eip712HashTypedData(eip712HashTypedDataCall),

}

Container for all the Vm function calls.

Variants

addr(addrCall)

dumpState(dumpStateCall)

getNonce_0(getNonce_0Call)

getNonce_1(getNonce_1Call)

load(loadCall)

loadAllocs(loadAllocsCall)

startDebugTraceRecording(startDebugTraceRecordingCall)

stopAndReturnDebugTraceRecording(stopAndReturnDebugTraceRecordingCall)

cloneAccount(cloneAccountCall)

record(recordCall)

stopRecord(stopRecordCall)

accesses(accessesCall)

startStateDiffRecording(startStateDiffRecordingCall)

stopAndReturnStateDiff(stopAndReturnStateDiffCall)

getStateDiff(getStateDiffCall)

getStateDiffJson(getStateDiffJsonCall)

getStorageAccesses(getStorageAccessesCall)

startMappingRecording(startMappingRecordingCall)

stopMappingRecording(stopMappingRecordingCall)

getMappingLength(getMappingLengthCall)

getMappingSlotAt(getMappingSlotAtCall)

getMappingKeyAndParentOf(getMappingKeyAndParentOfCall)

chainId(chainIdCall)

coinbase(coinbaseCall)

difficulty(difficultyCall)

fee(feeCall)

prevrandao_0(prevrandao_0Call)

prevrandao_1(prevrandao_1Call)

blobhashes(blobhashesCall)

getBlobhashes(getBlobhashesCall)

roll(rollCall)

getBlockNumber(getBlockNumberCall)

txGasPrice(txGasPriceCall)

warp(warpCall)

getBlockTimestamp(getBlockTimestampCall)

getRawBlockHeader(getRawBlockHeaderCall)

blobBaseFee(blobBaseFeeCall)

getBlobBaseFee(getBlobBaseFeeCall)

setBlockhash(setBlockhashCall)

deal(dealCall)

etch(etchCall)

resetNonce(resetNonceCall)

setNonce(setNonceCall)

setNonceUnsafe(setNonceUnsafeCall)

store(storeCall)

cool(coolCall)

accessList(accessListCall)

noAccessList(noAccessListCall)

warmSlot(warmSlotCall)

coolSlot(coolSlotCall)

clearMockedCalls(clearMockedCallsCall)

mockCall_0(mockCall_0Call)

mockCall_1(mockCall_1Call)

mockCall_2(mockCall_2Call)

mockCall_3(mockCall_3Call)

mockCalls_0(mockCalls_0Call)

mockCalls_1(mockCalls_1Call)

mockCallRevert_0(mockCallRevert_0Call)

mockCallRevert_1(mockCallRevert_1Call)

mockCallRevert_2(mockCallRevert_2Call)

mockCallRevert_3(mockCallRevert_3Call)

mockFunction(mockFunctionCall)

prank_0(prank_0Call)

startPrank_0(startPrank_0Call)

prank_1(prank_1Call)

startPrank_1(startPrank_1Call)

prank_2(prank_2Call)

startPrank_2(startPrank_2Call)

prank_3(prank_3Call)

startPrank_3(startPrank_3Call)

stopPrank(stopPrankCall)

readCallers(readCallersCall)

snapshotValue_0(snapshotValue_0Call)

snapshotValue_1(snapshotValue_1Call)

snapshotGasLastCall_0(snapshotGasLastCall_0Call)

snapshotGasLastCall_1(snapshotGasLastCall_1Call)

startSnapshotGas_0(startSnapshotGas_0Call)

startSnapshotGas_1(startSnapshotGas_1Call)

stopSnapshotGas_0(stopSnapshotGas_0Call)

stopSnapshotGas_1(stopSnapshotGas_1Call)

stopSnapshotGas_2(stopSnapshotGas_2Call)

snapshot(snapshotCall)

snapshotState(snapshotStateCall)

revertTo(revertToCall)

revertToState(revertToStateCall)

revertToAndDelete(revertToAndDeleteCall)

revertToStateAndDelete(revertToStateAndDeleteCall)

deleteSnapshot(deleteSnapshotCall)

deleteStateSnapshot(deleteStateSnapshotCall)

deleteSnapshots(deleteSnapshotsCall)

deleteStateSnapshots(deleteStateSnapshotsCall)

activeFork(activeForkCall)

createFork_0(createFork_0Call)

createFork_1(createFork_1Call)

createFork_2(createFork_2Call)

createSelectFork_0(createSelectFork_0Call)

createSelectFork_1(createSelectFork_1Call)

createSelectFork_2(createSelectFork_2Call)

rollFork_0(rollFork_0Call)

rollFork_1(rollFork_1Call)

rollFork_2(rollFork_2Call)

rollFork_3(rollFork_3Call)

selectFork(selectForkCall)

transact_0(transact_0Call)

transact_1(transact_1Call)

rpc_0(rpc_0Call)

rpc_1(rpc_1Call)

eth_getLogs(eth_getLogsCall)

allowCheatcodes(allowCheatcodesCall)

makePersistent_0(makePersistent_0Call)

makePersistent_1(makePersistent_1Call)

makePersistent_2(makePersistent_2Call)

makePersistent_3(makePersistent_3Call)

revokePersistent_0(revokePersistent_0Call)

revokePersistent_1(revokePersistent_1Call)

isPersistent(isPersistentCall)

recordLogs(recordLogsCall)

getRecordedLogs(getRecordedLogsCall)

pauseGasMetering(pauseGasMeteringCall)

resumeGasMetering(resumeGasMeteringCall)

resetGasMetering(resetGasMeteringCall)

lastCallGas(lastCallGasCall)

assume(assumeCall)

assumeNoRevert_0(assumeNoRevert_0Call)

assumeNoRevert_1(assumeNoRevert_1Call)

assumeNoRevert_2(assumeNoRevert_2Call)

breakpoint_0(breakpoint_0Call)

breakpoint_1(breakpoint_1Call)

getFoundryVersion(getFoundryVersionCall)

rpcUrl(rpcUrlCall)

rpcUrls(rpcUrlsCall)

rpcUrlStructs(rpcUrlStructsCall)

getChain_0(getChain_0Call)

getChain_1(getChain_1Call)

sleep(sleepCall)

expectCall_0(expectCall_0Call)

expectCall_1(expectCall_1Call)

expectCall_2(expectCall_2Call)

expectCall_3(expectCall_3Call)

expectCall_4(expectCall_4Call)

expectCall_5(expectCall_5Call)

expectCallMinGas_0(expectCallMinGas_0Call)

expectCallMinGas_1(expectCallMinGas_1Call)

expectEmit_0(expectEmit_0Call)

expectEmit_1(expectEmit_1Call)

expectEmit_2(expectEmit_2Call)

expectEmit_3(expectEmit_3Call)

expectEmit_4(expectEmit_4Call)

expectEmit_5(expectEmit_5Call)

expectEmit_6(expectEmit_6Call)

expectEmit_7(expectEmit_7Call)

expectEmitAnonymous_0(expectEmitAnonymous_0Call)

expectEmitAnonymous_1(expectEmitAnonymous_1Call)

expectEmitAnonymous_2(expectEmitAnonymous_2Call)

expectEmitAnonymous_3(expectEmitAnonymous_3Call)

expectCreate(expectCreateCall)

expectCreate2(expectCreate2Call)

expectRevert_0(expectRevert_0Call)

expectRevert_1(expectRevert_1Call)

expectRevert_2(expectRevert_2Call)

expectRevert_3(expectRevert_3Call)

expectRevert_4(expectRevert_4Call)

expectRevert_5(expectRevert_5Call)

expectRevert_6(expectRevert_6Call)

expectRevert_7(expectRevert_7Call)

expectRevert_8(expectRevert_8Call)

expectRevert_9(expectRevert_9Call)

expectRevert_10(expectRevert_10Call)

expectRevert_11(expectRevert_11Call)

expectPartialRevert_0(expectPartialRevert_0Call)

expectPartialRevert_1(expectPartialRevert_1Call)

_expectCheatcodeRevert_0(_expectCheatcodeRevert_0Call)

_expectCheatcodeRevert_1(_expectCheatcodeRevert_1Call)

_expectCheatcodeRevert_2(_expectCheatcodeRevert_2Call)

expectSafeMemory(expectSafeMemoryCall)

stopExpectSafeMemory(stopExpectSafeMemoryCall)

expectSafeMemoryCall(expectSafeMemoryCallCall)

skip_0(skip_0Call)

skip_1(skip_1Call)

assertTrue_0(assertTrue_0Call)

assertTrue_1(assertTrue_1Call)

assertFalse_0(assertFalse_0Call)

assertFalse_1(assertFalse_1Call)

assertEq_0(assertEq_0Call)

assertEq_1(assertEq_1Call)

assertEq_2(assertEq_2Call)

assertEq_3(assertEq_3Call)

assertEq_4(assertEq_4Call)

assertEq_5(assertEq_5Call)

assertEq_6(assertEq_6Call)

assertEq_7(assertEq_7Call)

assertEq_8(assertEq_8Call)

assertEq_9(assertEq_9Call)

assertEq_10(assertEq_10Call)

assertEq_11(assertEq_11Call)

assertEq_12(assertEq_12Call)

assertEq_13(assertEq_13Call)

assertEq_14(assertEq_14Call)

assertEq_15(assertEq_15Call)

assertEq_16(assertEq_16Call)

assertEq_17(assertEq_17Call)

assertEq_18(assertEq_18Call)

assertEq_19(assertEq_19Call)

assertEq_20(assertEq_20Call)

assertEq_21(assertEq_21Call)

assertEq_22(assertEq_22Call)

assertEq_23(assertEq_23Call)

assertEq_24(assertEq_24Call)

assertEq_25(assertEq_25Call)

assertEq_26(assertEq_26Call)

assertEq_27(assertEq_27Call)

assertEqDecimal_0(assertEqDecimal_0Call)

assertEqDecimal_1(assertEqDecimal_1Call)

assertEqDecimal_2(assertEqDecimal_2Call)

assertEqDecimal_3(assertEqDecimal_3Call)

assertNotEq_0(assertNotEq_0Call)

assertNotEq_1(assertNotEq_1Call)

assertNotEq_2(assertNotEq_2Call)

assertNotEq_3(assertNotEq_3Call)

assertNotEq_4(assertNotEq_4Call)

assertNotEq_5(assertNotEq_5Call)

assertNotEq_6(assertNotEq_6Call)

assertNotEq_7(assertNotEq_7Call)

assertNotEq_8(assertNotEq_8Call)

assertNotEq_9(assertNotEq_9Call)

assertNotEq_10(assertNotEq_10Call)

assertNotEq_11(assertNotEq_11Call)

assertNotEq_12(assertNotEq_12Call)

assertNotEq_13(assertNotEq_13Call)

assertNotEq_14(assertNotEq_14Call)

assertNotEq_15(assertNotEq_15Call)

assertNotEq_16(assertNotEq_16Call)

assertNotEq_17(assertNotEq_17Call)

assertNotEq_18(assertNotEq_18Call)

assertNotEq_19(assertNotEq_19Call)

assertNotEq_20(assertNotEq_20Call)

assertNotEq_21(assertNotEq_21Call)

assertNotEq_22(assertNotEq_22Call)

assertNotEq_23(assertNotEq_23Call)

assertNotEq_24(assertNotEq_24Call)

assertNotEq_25(assertNotEq_25Call)

assertNotEq_26(assertNotEq_26Call)

assertNotEq_27(assertNotEq_27Call)

assertNotEqDecimal_0(assertNotEqDecimal_0Call)

assertNotEqDecimal_1(assertNotEqDecimal_1Call)

assertNotEqDecimal_2(assertNotEqDecimal_2Call)

assertNotEqDecimal_3(assertNotEqDecimal_3Call)

assertGt_0(assertGt_0Call)

assertGt_1(assertGt_1Call)

assertGt_2(assertGt_2Call)

assertGt_3(assertGt_3Call)

assertGtDecimal_0(assertGtDecimal_0Call)

assertGtDecimal_1(assertGtDecimal_1Call)

assertGtDecimal_2(assertGtDecimal_2Call)

assertGtDecimal_3(assertGtDecimal_3Call)

assertGe_0(assertGe_0Call)

assertGe_1(assertGe_1Call)

assertGe_2(assertGe_2Call)

assertGe_3(assertGe_3Call)

assertGeDecimal_0(assertGeDecimal_0Call)

assertGeDecimal_1(assertGeDecimal_1Call)

assertGeDecimal_2(assertGeDecimal_2Call)

assertGeDecimal_3(assertGeDecimal_3Call)

assertLt_0(assertLt_0Call)

assertLt_1(assertLt_1Call)

assertLt_2(assertLt_2Call)

assertLt_3(assertLt_3Call)

assertLtDecimal_0(assertLtDecimal_0Call)

assertLtDecimal_1(assertLtDecimal_1Call)

assertLtDecimal_2(assertLtDecimal_2Call)

assertLtDecimal_3(assertLtDecimal_3Call)

assertLe_0(assertLe_0Call)

assertLe_1(assertLe_1Call)

assertLe_2(assertLe_2Call)

assertLe_3(assertLe_3Call)

assertLeDecimal_0(assertLeDecimal_0Call)

assertLeDecimal_1(assertLeDecimal_1Call)

assertLeDecimal_2(assertLeDecimal_2Call)

assertLeDecimal_3(assertLeDecimal_3Call)

assertApproxEqAbs_0(assertApproxEqAbs_0Call)

assertApproxEqAbs_1(assertApproxEqAbs_1Call)

assertApproxEqAbs_2(assertApproxEqAbs_2Call)

assertApproxEqAbs_3(assertApproxEqAbs_3Call)

assertApproxEqAbsDecimal_0(assertApproxEqAbsDecimal_0Call)

assertApproxEqAbsDecimal_1(assertApproxEqAbsDecimal_1Call)

assertApproxEqAbsDecimal_2(assertApproxEqAbsDecimal_2Call)

assertApproxEqAbsDecimal_3(assertApproxEqAbsDecimal_3Call)

assertApproxEqRel_0(assertApproxEqRel_0Call)

assertApproxEqRel_1(assertApproxEqRel_1Call)

assertApproxEqRel_2(assertApproxEqRel_2Call)

assertApproxEqRel_3(assertApproxEqRel_3Call)

assertApproxEqRelDecimal_0(assertApproxEqRelDecimal_0Call)

assertApproxEqRelDecimal_1(assertApproxEqRelDecimal_1Call)

assertApproxEqRelDecimal_2(assertApproxEqRelDecimal_2Call)

assertApproxEqRelDecimal_3(assertApproxEqRelDecimal_3Call)

foundryVersionAtLeast(foundryVersionAtLeastCall)

foundryVersionCmp(foundryVersionCmpCall)

exists(existsCall)

fsMetadata(fsMetadataCall)

isDir(isDirCall)

isFile(isFileCall)

projectRoot(projectRootCall)

unixTime(unixTimeCall)

closeFile(closeFileCall)

copyFile(copyFileCall)

createDir(createDirCall)

readDir_0(readDir_0Call)

readDir_1(readDir_1Call)

readDir_2(readDir_2Call)

readFile(readFileCall)

readFileBinary(readFileBinaryCall)

readLine(readLineCall)

readLink(readLinkCall)

removeDir(removeDirCall)

removeFile(removeFileCall)

writeFile(writeFileCall)

writeFileBinary(writeFileBinaryCall)

writeLine(writeLineCall)

getArtifactPathByCode(getArtifactPathByCodeCall)

getArtifactPathByDeployedCode(getArtifactPathByDeployedCodeCall)

getCode(getCodeCall)

deployCode_0(deployCode_0Call)

deployCode_1(deployCode_1Call)

deployCode_2(deployCode_2Call)

deployCode_3(deployCode_3Call)

deployCode_4(deployCode_4Call)

deployCode_5(deployCode_5Call)

deployCode_6(deployCode_6Call)

deployCode_7(deployCode_7Call)

getDeployedCode(getDeployedCodeCall)

getBroadcast(getBroadcastCall)

getBroadcasts_0(getBroadcasts_0Call)

getBroadcasts_1(getBroadcasts_1Call)

getDeployment_0(getDeployment_0Call)

getDeployment_1(getDeployment_1Call)

getDeployments(getDeploymentsCall)

ffi(ffiCall)

tryFfi(tryFfiCall)

prompt(promptCall)

promptSecret(promptSecretCall)

promptSecretUint(promptSecretUintCall)

promptAddress(promptAddressCall)

promptUint(promptUintCall)

setEnv(setEnvCall)

envExists(envExistsCall)

envBool_0(envBool_0Call)

envUint_0(envUint_0Call)

envInt_0(envInt_0Call)

envAddress_0(envAddress_0Call)

envBytes32_0(envBytes32_0Call)

envString_0(envString_0Call)

envBytes_0(envBytes_0Call)

envBool_1(envBool_1Call)

envUint_1(envUint_1Call)

envInt_1(envInt_1Call)

envAddress_1(envAddress_1Call)

envBytes32_1(envBytes32_1Call)

envString_1(envString_1Call)

envBytes_1(envBytes_1Call)

envOr_0(envOr_0Call)

envOr_1(envOr_1Call)

envOr_2(envOr_2Call)

envOr_3(envOr_3Call)

envOr_4(envOr_4Call)

envOr_5(envOr_5Call)

envOr_6(envOr_6Call)

envOr_7(envOr_7Call)

envOr_8(envOr_8Call)

envOr_9(envOr_9Call)

envOr_10(envOr_10Call)

envOr_11(envOr_11Call)

envOr_12(envOr_12Call)

envOr_13(envOr_13Call)

isContext(isContextCall)

forkChains(forkChainsCall)

forkChainIds(forkChainIdsCall)

forkChain(forkChainCall)

forkChainId(forkChainIdCall)

forkRpcUrl(forkRpcUrlCall)

forkChainRpcUrl(forkChainRpcUrlCall)

forkBool(forkBoolCall)

forkChainBool(forkChainBoolCall)

forkInt(forkIntCall)

forkChainInt(forkChainIntCall)

forkUint(forkUintCall)

forkChainUint(forkChainUintCall)

forkAddress(forkAddressCall)

forkChainAddress(forkChainAddressCall)

forkBytes32(forkBytes32Call)

forkChainBytes32(forkChainBytes32Call)

forkString(forkStringCall)

forkChainString(forkChainStringCall)

forkBytes(forkBytesCall)

forkChainBytes(forkChainBytesCall)

broadcast_0(broadcast_0Call)

broadcast_1(broadcast_1Call)

broadcast_2(broadcast_2Call)

startBroadcast_0(startBroadcast_0Call)

startBroadcast_1(startBroadcast_1Call)

startBroadcast_2(startBroadcast_2Call)

stopBroadcast(stopBroadcastCall)

broadcastRawTransaction(broadcastRawTransactionCall)

signDelegation_0(signDelegation_0Call)

signDelegation_1(signDelegation_1Call)

signDelegation_2(signDelegation_2Call)

attachDelegation_0(attachDelegation_0Call)

attachDelegation_1(attachDelegation_1Call)

signAndAttachDelegation_0(signAndAttachDelegation_0Call)

signAndAttachDelegation_1(signAndAttachDelegation_1Call)

signAndAttachDelegation_2(signAndAttachDelegation_2Call)

attachBlob(attachBlobCall)

getWallets(getWalletsCall)

toString_0(toString_0Call)

toString_1(toString_1Call)

toString_2(toString_2Call)

toString_3(toString_3Call)

toString_4(toString_4Call)

toString_5(toString_5Call)

parseBytes(parseBytesCall)

parseAddress(parseAddressCall)

parseUint(parseUintCall)

parseInt(parseIntCall)

parseBytes32(parseBytes32Call)

parseBool(parseBoolCall)

toLowercase(toLowercaseCall)

toUppercase(toUppercaseCall)

trim(trimCall)

replace(replaceCall)

split(splitCall)

indexOf(indexOfCall)

contains(containsCall)

keyExists(keyExistsCall)

keyExistsJson(keyExistsJsonCall)

parseJson_0(parseJson_0Call)

parseJson_1(parseJson_1Call)

parseJsonUint(parseJsonUintCall)

parseJsonUintArray(parseJsonUintArrayCall)

parseJsonInt(parseJsonIntCall)

parseJsonIntArray(parseJsonIntArrayCall)

parseJsonBool(parseJsonBoolCall)

parseJsonBoolArray(parseJsonBoolArrayCall)

parseJsonAddress(parseJsonAddressCall)

parseJsonAddressArray(parseJsonAddressArrayCall)

parseJsonString(parseJsonStringCall)

parseJsonStringArray(parseJsonStringArrayCall)

parseJsonBytes(parseJsonBytesCall)

parseJsonBytesArray(parseJsonBytesArrayCall)

parseJsonBytes32(parseJsonBytes32Call)

parseJsonBytes32Array(parseJsonBytes32ArrayCall)

parseJsonType_0(parseJsonType_0Call)

parseJsonType_1(parseJsonType_1Call)

parseJsonTypeArray(parseJsonTypeArrayCall)

parseJsonKeys(parseJsonKeysCall)

serializeJson(serializeJsonCall)

serializeBool_0(serializeBool_0Call)

serializeUint_0(serializeUint_0Call)

serializeUintToHex(serializeUintToHexCall)

serializeInt_0(serializeInt_0Call)

serializeAddress_0(serializeAddress_0Call)

serializeBytes32_0(serializeBytes32_0Call)

serializeString_0(serializeString_0Call)

serializeBytes_0(serializeBytes_0Call)

serializeBool_1(serializeBool_1Call)

serializeUint_1(serializeUint_1Call)

serializeInt_1(serializeInt_1Call)

serializeAddress_1(serializeAddress_1Call)

serializeBytes32_1(serializeBytes32_1Call)

serializeString_1(serializeString_1Call)

serializeBytes_1(serializeBytes_1Call)

serializeJsonType_0(serializeJsonType_0Call)

serializeJsonType_1(serializeJsonType_1Call)

writeJson_0(writeJson_0Call)

writeJson_1(writeJson_1Call)

keyExistsToml(keyExistsTomlCall)

parseToml_0(parseToml_0Call)

parseToml_1(parseToml_1Call)

parseTomlUint(parseTomlUintCall)

parseTomlUintArray(parseTomlUintArrayCall)

parseTomlInt(parseTomlIntCall)

parseTomlIntArray(parseTomlIntArrayCall)

parseTomlBool(parseTomlBoolCall)

parseTomlBoolArray(parseTomlBoolArrayCall)

parseTomlAddress(parseTomlAddressCall)

parseTomlAddressArray(parseTomlAddressArrayCall)

parseTomlString(parseTomlStringCall)

parseTomlStringArray(parseTomlStringArrayCall)

parseTomlBytes(parseTomlBytesCall)

parseTomlBytesArray(parseTomlBytesArrayCall)

parseTomlBytes32(parseTomlBytes32Call)

parseTomlBytes32Array(parseTomlBytes32ArrayCall)

parseTomlType_0(parseTomlType_0Call)

parseTomlType_1(parseTomlType_1Call)

parseTomlTypeArray(parseTomlTypeArrayCall)

parseTomlKeys(parseTomlKeysCall)

writeToml_0(writeToml_0Call)

writeToml_1(writeToml_1Call)

createWallet_0(createWallet_0Call)

createWallet_1(createWallet_1Call)

createWallet_2(createWallet_2Call)

sign_0(sign_0Call)

signCompact_0(signCompact_0Call)

sign_1(sign_1Call)

signCompact_1(signCompact_1Call)

sign_2(sign_2Call)

signCompact_2(signCompact_2Call)

sign_3(sign_3Call)

signCompact_3(signCompact_3Call)

signP256(signP256Call)

publicKeyP256(publicKeyP256Call)

deriveKey_0(deriveKey_0Call)

deriveKey_1(deriveKey_1Call)

deriveKey_2(deriveKey_2Call)

deriveKey_3(deriveKey_3Call)

rememberKey(rememberKeyCall)

rememberKeys_0(rememberKeys_0Call)

rememberKeys_1(rememberKeys_1Call)

label(labelCall)

getLabel(getLabelCall)

computeCreateAddress(computeCreateAddressCall)

computeCreate2Address_0(computeCreate2Address_0Call)

computeCreate2Address_1(computeCreate2Address_1Call)

toBase64_0(toBase64_0Call)

toBase64_1(toBase64_1Call)

toBase64URL_0(toBase64URL_0Call)

toBase64URL_1(toBase64URL_1Call)

ensNamehash(ensNamehashCall)

randomUint_0(randomUint_0Call)

randomUint_1(randomUint_1Call)

randomUint_2(randomUint_2Call)

randomAddress(randomAddressCall)

randomInt_0(randomInt_0Call)

randomInt_1(randomInt_1Call)

randomBool(randomBoolCall)

randomBytes(randomBytesCall)

randomBytes4(randomBytes4Call)

randomBytes8(randomBytes8Call)

pauseTracing(pauseTracingCall)

resumeTracing(resumeTracingCall)

copyStorage(copyStorageCall)

setArbitraryStorage_0(setArbitraryStorage_0Call)

setArbitraryStorage_1(setArbitraryStorage_1Call)

sort(sortCall)

shuffle(shuffleCall)

setSeed(setSeedCall)

interceptInitcode(interceptInitcodeCall)

eip712HashType_0(eip712HashType_0Call)

eip712HashType_1(eip712HashType_1Call)

eip712HashStruct_0(eip712HashStruct_0Call)

eip712HashStruct_1(eip712HashStruct_1Call)

eip712HashTypedData(eip712HashTypedDataCall)

Implementations

impl VmCalls

pub const SELECTORS: &'static [[u8; 4]]

All the selectors of this enum.

Note that the selectors might not be in the same order as the variants. No guarantees are made about the order of the selectors.

Prefer using SolInterface methods instead.