Module forge::multi_runner

Forge test runner for multiple contracts.

Structs

MultiContractRunner

A multi contract runner receives a set of contracts deployed in an EVM instance and proceeds to run all test functions in these contracts.

MultiContractRunnerBuilder

Builder used for instantiating the multi-contract runner

Functions

is_matching_test 🔒

Returns true if the function is a test function that matches the given filter.

matches_contract 🔒

Type Aliases

DeployableContracts