forge
Module progress

Structs

TestsProgress

Clonable wrapper around TestsProgressState.

TestsProgressState

State of [ProgressBar]s displayed for the given test run. Shows progress of all test suites matching filter. For each test within the test suite an individual progress bar is displayed. When a test suite completes, their progress is removed from overall progress and result summary is displayed.

Functions

start_fuzz_progress

Helper function for creating fuzz test progress bar.