forge::revm::bytecode::bitvec::boxed

Struct BitBox

#[repr(transparent)]
pub struct BitBox<T = usize, O = Lsb0>
where
    T: BitStore,
    O: BitOrder,
{
    bitspan: BitSpan<Mut, T, O>,
}

Fixed-Size, Heap-Allocated, Bit Slice

BitBox is a heap-allocated BitSlice region. It is a distinct type because the implementation of bit-slice pointers means that Box<BitSlice> cannot exist. It can be created by cloning a bit-slice into the heap, or by freezing the allocation of a BitVec

Original

Box<[T]>

API Differences

As with BitSlice, this takes a pair of BitOrder and BitStore type parameters to govern the buffer’s memory representation. Because BitSlice is unsized, BitBox has almost none of the Box API, and is difficult to use directly.

Behavior

BitBox, like &BitSlice, is an opaque pointer to a bit-addressed slice region. Unlike &BitSlice, it uses the allocator to guarantee that it is the sole accessor to the referent buffer, and is able to use that uniqueness guarantee to specialize some BitSlice behavior to be faster or more efficient.

Safety

BitBox is, essentially, a NonNull<BitSlice<T, O>> pointer. The internal value is opaque and cannot be inspected or modified by user code.

If you attempt to do so, your program becomes inconsistent. You will likely break the allocator’s internal state and cause a crash. No guarantees of crash or recovery are provided. Do not inspect or modify the BitBox handle value.

Construction

The simplest way to construct a BitBox is by using the bitbox! macro. You can also explicitly clone a BitSlice with BitBox::from_bitslice, or freeze a BitVec with BitVec::into_boxed_bitslice.

Examples

use bitvec::prelude::*;

let a = BitBox::from_bitslice(bits![1, 0, 1, 1, 0]);
let b = bitbox![0, 1, 0, 0, 1];

let b_raw: *mut BitSlice = BitBox::into_raw(b);
let b_reformed = unsafe { BitBox::from_raw(b_raw) };

Fields
bitspan: BitSpan<Mut, T, O>
Implementations

impl<T, O> BitBox<T, O>
where T: BitStore, O: BitOrder,

pub unsafe fn from_raw(raw: *mut BitSlice<T, O>) -> BitBox<T, O>

Constructs a bit-box from a raw bit-slice pointer.

This converts a *mut BitSlice pointer that had previously been produced by either ::into_raw() or ::leak() and restores the bit-box containing it.

Original

Box::from_raw

Safety

You must only call this function on pointers produced by leaking a prior BitBox; you may not modify the value of a pointer returned by ::into_raw(), nor may you conjure pointer values of your own. Doing so will corrupt the allocator state.

You must only call this function on any given leaked pointer at most once. Not calling it at all will merely render the allocated memory unreachable for the duration of the program runtime, a normal (and safe) memory leak. Calling it once restores ordinary functionality, and ensures ordinary destruction at or before program termination. However, calling it more than once on the same pointer will introduce data races, use-after-free, and/or double-free errors.

Examples

use bitvec::prelude::*;

let bb = bitbox![0; 80];
let ptr: *mut BitSlice = BitBox::into_raw(bb);
let bb = unsafe { BitBox::from_raw(ptr) };
// unsafe { BitBox::from_raw(ptr) }; // UAF crash!

pub fn into_raw(this: BitBox<T, O>) -> *mut BitSlice<T, O>

Consumes the bit-box, returning a raw bit-slice pointer.

Bit-slice pointers are always correctly encoded and non-null. The referent region is dereferenceäble *as a BitSlice for the remainder of the program, or until it is first passed to ::from_raw(), whichever comes first. Once the pointer is first passed to ::from_raw(), all copies of that pointer become invalid to dereference.

Original

Box::into_raw

Examples

use bitvec::prelude::*;

let bb = bitbox![0; 80];
let ptr = BitBox::into_raw(bb);
let bb = unsafe { BitBox::from_raw(ptr) };

You may not deällocate pointers produced by this function through any other means.

pub fn leak<'a>(this: BitBox<T, O>) -> &'a mut BitSlice<T, O>
where T: 'a,

Deliberately leaks the allocated memory, returning an &'static mut BitSlice reference.

This differs from ::into_raw() in that the reference is safe to use and can be tracked by the Rust borrow-checking system. Like the bit-slice pointer produced by ::into_raw(), this reference can be un-leaked by passing it into ::from_raw() to reclaim the memory.

Original

Box::leak

Examples

use bitvec::prelude::*;

let bb = bitbox![0; 80];
let static_ref: &'static mut BitSlice = BitBox::leak(bb);

static_ref.set(0, true);
assert!(static_ref[0]);
let _ = unsafe {
  BitBox::from_raw(static_ref)
};

impl<T, O> BitBox<T, O>
where T: BitStore, O: BitOrder,

pub fn from_bitslice(slice: &BitSlice<T, O>) -> BitBox<T, O>

Copies a bit-slice region into a new bit-box allocation.

The referent memory is memcpyd into the heap, exactly preserving the original bit-slice’s memory layout and contents. This allows the function to run as fast as possible, but misaligned source bit-slices may result in decreased performance or unexpected layout behavior during use. You can use .force_align() to ensure that the referent bit-slice is aligned in memory.

Notes

Bits in the allocation of the source bit-slice, but outside its own description of that memory, have an unspecified, but initialized, value. You may not rely on their contents in any way, and you should call .force_align() and/or .fill_uninitialized() if you are going to inspect the underlying memory of the new allocation.

Examples

use bitvec::prelude::*;

let data = 0b0101_1011u8;
let bits = data.view_bits::<Msb0>();
let bb = BitBox::from_bitslice(&bits[2 ..]);
assert_eq!(bb, bits[2 ..]);

pub fn from_boxed_slice(boxed: Box<[T]>) -> BitBox<T, O>

Converts a Box<[T]> into a BitBox<T, O>, in place.

This does not affect the referent buffer, and only transforms the handle.

Panics

This panics if the provided boxed slice is too long to view as a bit-slice region.

Examples

use bitvec::prelude::*;

let boxed: Box<[u8]> = Box::new([0; 40]);
let addr = boxed.as_ptr();
let bb = BitBox::<u8>::from_boxed_slice(boxed);
assert_eq!(bb, bits![0; 320]);
assert_eq!(addr, bb.as_raw_slice().as_ptr());

pub fn try_from_boxed_slice(boxed: Box<[T]>) -> Result<BitBox<T, O>, Box<[T]>>

Attempts to convert an ordinary boxed slice into a boxed bit-slice.

This does not perform a copy or reällocation; it only attempts to transform the handle. Because Box<[T]> can be longer than BitBoxes, it may fail, and will return the original handle if it does.

It is unlikely that you have a single Box<[_]> that is too large to convert into a bit-box. You can find the length restrictions as the bit-slice associated constants MAX_BITS and MAX_ELTS.

Examples

use bitvec::prelude::*;

let boxed: Box<[u8]> = Box::new([0u8; 40]);
let addr = boxed.as_ptr();
let bb = BitBox::<u8>::try_from_boxed_slice(boxed).unwrap();
assert_eq!(bb, bits![0; 320]);
assert_eq!(addr, bb.as_raw_slice().as_ptr());

pub fn into_boxed_slice(self) -> Box<[T]>

Converts the bit-box back into an ordinary boxed element slice.

This does not touch the allocator or the buffer contents; it is purely a handle transform.

Examples

use bitvec::prelude::*;

let bb = bitbox![0; 5];
let addr = bb.as_raw_slice().as_ptr();
let boxed = bb.into_boxed_slice();
assert_eq!(boxed[..], [0][..]);
assert_eq!(addr, boxed.as_ptr());

pub fn into_bitvec(self) -> BitVec<T, O>

Converts the bit-box into a bit-vector.

This uses the Rust allocator API, and does not guarantee whether or not a reällocation occurs internally.

The resulting bit-vector can be converted back into a bit-box via BitBox::into_boxed_bitslice.

Original

slice::into_vec

API Differences

The original function is implemented in an impl<T> [T] block, despite taking a Box<[T]> receiver. Since BitBox cannot be used as an explicit receiver outside its own impl blocks, the method is relocated here.

Examples

use bitvec::prelude::*;

let bb = bitbox![0, 1, 0, 0, 1];
let bv = bb.into_bitvec();

assert_eq!(bv, bitvec![0, 1, 0, 0, 1]);

pub fn as_bitslice(&self) -> &BitSlice<T, O>

Explicitly views the bit-box as a bit-slice.

pub fn as_mut_bitslice(&mut self) -> &mut BitSlice<T, O>

Explicitly views the bit-box as a mutable bit-slice.

pub fn as_raw_slice(&self) -> &[T]

Views the bit-box as a slice of its underlying memory elements.

Because bit-boxes uniquely own their buffer, they can safely view the underlying buffer without dealing with contending neighbors.

pub fn as_raw_mut_slice(&mut self) -> &mut [T]

Views the bit-box as a mutable slice of its underlying memory elements.

Because bit-boxes uniquely own their buffer, they can safely view the underlying buffer without dealing with contending neighbors.

pub fn fill_uninitialized(&mut self, value: bool)

Sets the unused bits outside the BitBox buffer to a fixed value.

This method modifies all bits that the allocated buffer owns but which are outside the self.as_bitslice() view. bitvec guarantees that all owned bits are initialized to some value, but does not guarantee which value. This method can be used to make all such unused bits have a known value after the call, so that viewing the underlying memory directly has consistent results.

Note that the crate implementation guarantees that all bits owned by its handles are stably initialized according to the language and compiler rules! bitvec will never cause UB by using uninitialized memory.

Examples

use bitvec::prelude::*;

let bits = 0b1011_0101u8.view_bits::<Msb0>();
let mut bb = BitBox::from_bitslice(&bits[2 .. 6]);
assert_eq!(bb.count_ones(), 3);
// Remember, the two bits on each edge are unspecified, and cannot be
// observed! They must be masked away for the test to be meaningful.
assert_eq!(bb.as_raw_slice()[0] & 0x3C, 0b00_1101_00u8);

bb.fill_uninitialized(false);
assert_eq!(bb.as_raw_slice(), &[0b00_1101_00u8]);

bb.fill_uninitialized(true);
assert_eq!(bb.as_raw_slice(), &[0b11_1101_11u8]);

pub fn force_align(&mut self)

Ensures that the allocated buffer has no dead bits between the start of the buffer and the start of the live bit-slice.

This is useful for ensuring a consistent memory layout in bit-boxes created by cloning an arbitrary bit-slice into the heap. As bit-slices can begin and end anywhere in memory, the ::from_bitslice() function does not attempt to normalize them and only does a fast element-wise copy when creating the bit-box.

The value of dead bits that are in the allocation but not in the live region are initialized, but do not have a specified value. After calling this method, you should use .fill_uninitialized() to set the excess bits in the buffer to a fixed value.

Examples

use bitvec::prelude::*;

let bits = &0b10_1101_01u8.view_bits::<Msb0>()[2 .. 6];
let mut bb = BitBox::from_bitslice(bits);
// Remember, the two bits on each edge are unspecified, and cannot be
// observed! They must be masked away for the test to be meaningful.
assert_eq!(bb.as_raw_slice()[0] & 0x3C, 0b00_1101_00u8);

bb.force_align();
bb.fill_uninitialized(false);
assert_eq!(bb.as_raw_slice(), &[0b1101_0000u8]);