forge::revm

Crate context

EVM execution context.

Modules

block

This module contains BlockEnv and it implements Block trait.

cfg

This module contains CfgEnv and implements Cfg trait for it.

context

This module contains Context struct and implements ContextTr trait for it.

either

The enum Either with variants Left and Right is a general purpose sum type with two cases.

evm

This module contains Evm struct.

host

Host interface for external blockchain state access.

journal

This module contains Journal struct and implements JournalTr trait for it.

journaled_state

Journaled state trait JournalTr and related types.

local

Local context that is filled by execution.

result

Result of the EVM execution. Containing both execution result, state and errors.

transaction

Transaction trait Transaction and associated types.

tx

This module contains TxEnv struct and implements Transaction trait for it.

Structs

BlockEnv

The block environment

CfgEnv

EVM configuration

Context

EVM context contains data that EVM needs for execution.

DummyHost

Dummy host that implements Host trait and returns all default values.

Evm

Main EVM structure that contains all data needed for execution.

FrameStack

Non-empty, item-pooling Vec.

FrameToken

Used to guarantee that a frame is initialized before use.

Journal

A journal of state changes internal to the EVM

JournalInner

Inner journal state that contains journal and state changes.

LocalContext

Local context that is filled by execution.

OutFrame

A potentially initialized frame. Used when initializing a new frame in the main loop.

TxEnv

The Transaction Environment is a struct that contains all fields that can be found in all Ethereum transaction, including EIP-4844, EIP-7702, EIP-7873, etc. It implements the Transaction trait, which is used inside the EVM to execute a transaction.

Enums

ContextError

Inner Context error used for Interpreter to set error without returning it from instruction

CreateScheme

Create scheme

JournalEntry

Journal entries that are used to track changes to the state and are used to revert it.

TransactionType

Transaction types of all Ethereum transactions

Traits

Block

Trait for retrieving block information required for execution.

Cfg

Configuration for the EVM.

ContextSetters

Trait for setting the transaction and block in the context.

ContextTr

Trait that defines the context of the EVM execution.

DBErrorMarker

Database error marker is needed to implement From conversion for Error type.

Database

EVM database interface.

Host

Host trait with all methods that are needed by the Interpreter.

JournalEntryTr

Trait for tracking and reverting state changes in the EVM. Journal entry contains information about state changes that can be reverted.

JournalTr

Trait that contains database and journal of all changes that were made to the state.

LocalContextTr

Local context used for caching initcode from Initcode transactions.

Transaction

Main Transaction trait that abstracts and specifies all transaction currently supported by Ethereum

Type Aliases

TransactTo

Transaction destination