forge::revm::context

Crate either

The enum Either with variants Left and Right is a general purpose sum type with two cases.

Crate features:

"std" Enabled by default. Disable to make the library #![no_std].

"serde" Disabled by default. Enable to #[derive(Serialize, Deserialize)] for Either

Modules

serde_untagged

Untagged serialization/deserialization support for Either<L, R>.

serde_untagged_optional

Untagged serialization/deserialization support for Option<Either<L, R>>.

Macros

for_both

Evaluate the provided expression for both Either::Left and Either::Right.

try_left

Macro for unwrapping the left side of an Either, which fails early with the opposite side. Can only be used in functions that return Either because of the early return of Right that it provides.

try_right

Dual to try_left!, see its documentation for more information.

Structs

IterEither

Iterator that maps left or right iterators to corresponding Either-wrapped items.

Enums

Either

The enum Either with variants Left and Right is a general purpose sum type with two cases.

Traits

IntoEither

Provides methods for converting a type Self into either a Left or Right variant of Either<Self, Self>.