forge::revm::context::evm

Struct Evm

pub struct Evm<CTX, INSP, I, P, F> {
    pub ctx: CTX,
    pub inspector: INSP,
    pub instruction: I,
    pub precompiles: P,
    pub frame_stack: FrameStack<F>,
}

Main EVM structure that contains all data needed for execution.

Fields
ctx: CTX

context_interface::ContextTr of the EVM it is used to fetch data from database.

inspector: INSP

Inspector of the EVM it is used to inspect the EVM. Its trait are defined in revm-inspector crate.

instruction: I

Instructions provider of the EVM it is used to execute instructions. InstructionProvider trait is defined in revm-handler crate.

precompiles: P

Precompile provider of the EVM it is used to execute precompiles. PrecompileProvider trait is defined in revm-handler crate.

frame_stack: FrameStack<F>

Frame that is going to be executed.

Implementations

impl<CTX, I, P, F> Evm<CTX, (), I, P, F>

pub fn new(ctx: CTX, instruction: I, precompiles: P) -> Evm<CTX, (), I, P, F>

Create a new EVM instance with a given context, instruction set, and precompile provider.

Inspector will be set to ().

impl<CTX, I, INSP, P, F> Evm<CTX, INSP, I, P, F>

pub fn new_with_inspector( ctx: CTX, inspector: INSP, instruction: I, precompiles: P, ) -> Evm<CTX, INSP, I, P, F>

Create a new EVM instance with a given context, inspector, instruction set, and precompile provider.

impl<CTX, INSP, I, P, F> Evm<CTX, INSP, I, P, F>

pub fn with_inspector<OINSP>(self, inspector: OINSP) -> Evm<CTX, OINSP, I, P, F>

Consumed self and returns new Evm type with given Inspector.

pub fn with_precompiles<OP>(self, precompiles: OP) -> Evm<CTX, INSP, I, OP, F>

Consumes self and returns new Evm type with given Precompiles.

pub fn into_inspector(self) -> INSP

Consumes self and returns inner Inspector.