forge::revm

Crate context_interface

EVM execution context interface.

Modules

block

Block related types and functions.

cfg

Configuration for the EVM. Containing SpecId.

context

Context trait and related types.

local

Local context trait LocalContextTr and related types.

Structs

DummyHost

Dummy host that implements Host trait and returns all default values.

FrameStack

Non-empty, item-pooling Vec.

FrameToken

Used to guarantee that a frame is initialized before use.

OutFrame

A potentially initialized frame. Used when initializing a new frame in the main loop.

Enums

ContextError

Inner Context error used for Interpreter to set error without returning it from instruction

CreateScheme

Create scheme

TransactionType

Transaction types of all Ethereum transactions

Traits

Block

Trait for retrieving block information required for execution.

Cfg

Configuration for the EVM.

ContextSetters

Trait for setting the transaction and block in the context.

ContextTr

Trait that defines the context of the EVM execution.

DBErrorMarker

Database error marker is needed to implement From conversion for Error type.

Database

EVM database interface.

Host

Host trait with all methods that are needed by the Interpreter.

JournalTr

Trait that contains database and journal of all changes that were made to the state.

LocalContextTr

Local context used for caching initcode from Initcode transactions.

Transaction

Main Transaction trait that abstracts and specifies all transaction currently supported by Ethereum

Type Aliases

TransactTo

Transaction destination