forge::revm::context_interface::block

Module blob

Blob (EIP-4844) related functions and types. BlobExcessGasAndPrice is struct that helps with calculating blob gas price and excess blob gas.

See also the EIP-4844 helpers.

calc_blob_gasprice and calc_excess_blob_gas are used to calculate the blob gas price and excess blob gas.

BlobExcessGasAndPrice is used to store the blob gas price and excess blob gas.s

Structs

BlobExcessGasAndPrice

Structure holding block blob excess gas and it calculates blob fee

Functions

calc_blob_gasprice

Calculates the blob gas price from the header’s excess blob gas field.

calc_excess_blob_gas

Calculates the excess_blob_gas from the parent header’s blob_gas_used and excess_blob_gas. Uses calc_excess_blob_gas_osaka internally.

calc_excess_blob_gas_osaka

Calculates the excess_blob_gas from the parent header’s blob_gas_used and excess_blob_gas.

fake_exponential

Approximates factor * e ** (numerator / denominator) using Taylor expansion.

get_base_fee_per_blob_gas

Calculates the base fee per blob gas. Calls calc_blob_gasprice internally. Name of the function is aligned with EIP-4844 spec.