forge::revm::database

Module empty_db

Empty database implementation.

Structs

EmptyDBTyped

An empty database that always returns default values when queried

Type Aliases

EmptyDB

An empty database that always returns default values when queried