forge::revm::database

Module in_memory_db

Structs

BenchmarkDB

Custom benchmarking DB that only has account info for the zero address.

Cache

A cache used in CacheDB. Its kept separate so it can be used independently.

CacheDB

A Database implementation that stores all state changes in memory.

DbAccount

Enums

AccountState

Constants

BENCH_CALLER

BENCH_CALLER_BALANCE

BENCH_TARGET

BENCH_TARGET_BALANCE

EEADDRESS

CALLER address

FFADDRESS

BYTECODE address

Type Aliases

InMemoryDB

A Database implementation that stores all state changes in memory.