forge::revm::database::states

Module bundle_state

Bundle state management.

Structs

BundleBuilder

This builder is used to help to facilitate the initialization of BundleState struct

BundleState

Bundle state contain only values that got changed

Enums

BundleRetention

Bundle retention policy for applying substate to the bundle.

OriginalValuesKnown

Option for BundleState when converting it to the plain state.