forge::revm

Crate database_interface

Database interface.

Structs

ArcUpgradeError

Error type for implementation of TryDatabaseCommit on Arc.

EmptyDBTyped

An empty database that always returns default values when queried

WrapDatabaseRef

Wraps a DatabaseRef to provide a Database implementation.

Constants

BENCH_CALLER

BENCH_CALLER address

BENCH_CALLER_BALANCE

BENCH_CALLER_BALANCE balance

BENCH_TARGET

BENCH_TARGET address

BENCH_TARGET_BALANCE

BENCH_TARGET_BALANCE balance

EEADDRESS

Address with all 0xee..ee in it. Used for testing.

FFADDRESS

Address with all 0xff..ff in it. Used for testing.

Traits

DBErrorMarker

Database error marker is needed to implement From conversion for Error type.

Database

EVM database interface.

DatabaseCommit

EVM database commit interface.

DatabaseRef

EVM database interface.

TryDatabaseCommit

EVM database commit interface that can fail.

Type Aliases

EmptyDB

An empty database that always returns default values when queried