forge::revm::db
Module states

Modules

account_status

bundle_account

bundle_state

cache

cache_account

changes

plain_account

reverts

state

state_builder

transition_account

transition_state

Structs

AccountRevert

Assumption is that Revert can return full state from any future state to any past state.

BundleAccount

Account information focused on creating of database changesets and Reverts.

BundleBuilder

This builder is used to help to facilitate the initialization of BundleState struct

BundleState

Bundle state contain only values that got changed

CacheAccount

Cache account contains plain state that gets updated at every transaction when evm output is applied to CacheState.

CacheState

Cache state contains both modified and original values.

PlainAccount

PlainStateReverts

Plain state reverts are used to easily store reverts into database.

PlainStorageChangeset

Plain storage changeset. Used to apply storage changes of plain state to the database.

PlainStorageRevert

Plain Storage Revert. Containing old values of changed storage.

State

State of blockchain.

StateBuilder

Allows building of State and initializing it with different options.

StateChangeset

accounts/storages/contracts for inclusion into database. Structure is made so it is easier to apply directly to database that mostly have separate tables to store account/storage/contract data.

StorageSlot

This type keeps track of the current value of a storage slot.

TransitionAccount

Account Created when EVM state is merged to cache state. And it is sent to Block state.

TransitionState

Enums

AccountStatus

AccountStatus represents the various states an account can be in after being loaded from the database.

OriginalValuesKnown

Option for BundleState when converting it to the plain state.

RevertToSlot

So storage can have multiple types:

Type Aliases

DBBox

Database boxed with a lifetime and Send.

StateDBBox

More constrained version of State that uses Boxed database with a lifetime.

StorageWithOriginalValues

This storage represent values that are before block changed.