forge::revm

Module handler

Modules

execution

generic

mainnet

Mainnet related handlers.

post_execution

pre_execution

register

validation

Structs

ExecutionHandler

Handles related to stack frames.

Handler

Handler acts as a proxy and allow to define different behavior for different sections of the code. This allows nice integration of different chains or to disable some mainnet behavior.

PostExecutionHandler

Handles related to post execution after the stack loop is finished.

PreExecutionHandler

Handles related to pre execution before the stack loop is started.

ValidationHandler

Handles related to validation.

Type Aliases

DeductCallerHandle

Deduct the caller to its limit.

EndHandle

End handle, takes result and state and returns final result. This will be called after all the other handlers.

FrameCallHandle

Handle sub call.

FrameCallReturnHandle

Handle call return

FrameCreateHandle

Handle sub create.

FrameCreateReturnHandle

Handle create return

GenericContextHandle

Generic Handle that takes a mutable reference to the context and returns a result.

GenericContextHandleRet

Generic handle that takes a mutable reference to the context and returns a result.

InsertCallOutcomeHandle

Insert call outcome to the parent

InsertCreateOutcomeHandle

Insert call outcome to the parent

LoadAccountsHandle

Load access list accounts and beneficiary. There is no need to load Caller as it is assumed that it will be loaded in DeductCallerHandle.

LoadPrecompilesHandle

Loads precompiles into Evm

OutputHandle

Main return handle, takes state from journal and transforms internal result to external.

ReimburseCallerHandle

Reimburse the caller with ethereum it didn’t spent.

RewardBeneficiaryHandle

Reward beneficiary with transaction rewards.

ValidateEnvHandle

Handle that validates env.

ValidateInitialTxGasHandle

Initial gas calculation handle

ValidateTxEnvAgainstState

Handle that validates transaction environment against the state. Second parametar is initial gas.