forge::revm::handler

Module evm

Core EVM traits for execution and frame management.

Traits

EvmTr

A trait that integrates context, instruction set, and precompiles to create an EVM struct.

FrameTr

Trait for defining a frame type used in EVM execution.

Type Aliases

ContextDbError

Type alias for database error within a context

ContextTrDbError

Type alias for frame error within a context

FrameInitResult

Type alias for frame init result