forge::revm::handler
Module mainnet

Mainnet related handlers.

Functions

apply_eip7702_auth_list

Apply EIP-7702 auth list and return number gas refund on already created accounts.

call

Handle frame sub call.

call_return

clear

Clear handle clears error and journal state.

create

Handle frame sub create.

create_return

deduct_caller

Deducts the caller balance to the transaction limit.

deduct_caller_inner

Helper function that deducts the caller balance.

end

Mainnet end handle does not change the output.

eofcreate

Handle frame sub create.

eofcreate_return

execute_frame

Execute frame

insert_call_outcome

insert_create_outcome

insert_eofcreate_outcome

last_frame_return

Handle output of the transaction

load_accounts

Main load handle

load_precompiles

Main precompile load

output

Main return handle, returns the output of the transaction.

refund

reimburse_caller

reward_beneficiary

Reward beneficiary with gas fee.

validate_env

Validate environment for the mainnet.

validate_initial_tx_gas

Validate initial transaction gas.

validate_tx_against_state

Validates transaction against the state.