Module forge::revm::handler::post_execution

Structs

PostExecutionHandler

Handles related to post execution after the stack loop is finished.

Type Aliases

EndHandle

End handle, takes result and state and returns final result. This will be called after all the other handlers.

OutputHandle

Main return handle, takes state from journal and transforms internal result to external.

ReimburseCallerHandle

Reimburse the caller with ethereum it didn’t spent.

RewardBeneficiaryHandle

Reward beneficiary with transaction rewards.