forge::revm::handler

Module pre_execution

Handles related to the main function of the EVM.

They handle initial setup of the EVM, call loop and the final return of the EVM

Functions

apply_eip7702_auth_list

Apply EIP-7702 auth list and return number gas refund on already created accounts.

load_accounts

Loads and warms accounts for execution, including precompiles and access list.

validate_account_nonce_and_code

Validates caller account nonce and code according to EIP-3607.

validate_against_state_and_deduct_caller

Validates caller state and deducts transaction costs from the caller’s balance.