forge::revm::handler

Type Alias MainnetContext

pub type MainnetContext<DB> = Context<BlockEnv, TxEnv, CfgEnv, DB>;

Type alias for a mainnet context with standard Ethereum environment types.

Aliased Type

pub struct MainnetContext<DB> {
    pub block: BlockEnv,
    pub tx: TxEnv,
    pub cfg: CfgEnv,
    pub journaled_state: Journal<DB>,
    pub chain: (),
    pub local: LocalContext,
    pub error: Result<(), ContextError<<DB as Database>::Error>>,
}

Fields
block: BlockEnv

Block information.

tx: TxEnv

Transaction information.

cfg: CfgEnv

Configurations.

journaled_state: Journal<DB>

EVM State with journaling support and database.

chain: ()

Inner context.

local: LocalContext

Local context that is filled by execution.

error: Result<(), ContextError<<DB as Database>::Error>>

Error that happened during execution.

Layout

Note: Unable to compute type layout, possibly due to this type having generic parameters. Layout can only be computed for concrete, fully-instantiated types.