forge::revm

Crate inspector

Inspector is a crate that provides a set of traits that allow inspecting the EVM execution.

It is used to implement tracers that can be used to inspect the EVM execution. Implementing inspection is optional and it does not effect the core execution.

Modules

handler

Handler implementations for inspector integration.

inspectors

Inspector implementations.

Structs

CountInspector

Inspector that counts all opcodes that were called during execution.

NoOpInspector

Dummy Inspector, helpful as standalone replacement.

Traits

InspectCommitEvm

InspectCommitEvm is a API that allows inspecting similar to InspectEvm but it has functions that commit the state diff to the database.

InspectEvm

InspectEvm is a API that allows inspecting the EVM.

InspectSystemCallEvm

InspectSystemCallEvm is an API that allows inspecting system calls in the EVM.

Inspector

EVM hooks into execution.

InspectorEvmTr

Inspector EVM trait. Extends the EvmTr trait with inspector related methods.

InspectorFrame

Trait that extends the FrameTr trait with additional functionality that is needed for inspection.

InspectorHandler

Trait that extends Handler with inspection functionality.

JournalExt

Extends the journal with additional methods that are used by the inspector.

Functions

inspect_instructions

Run Interpreter loop with inspection support.