forge::revm::inspector::inspectors

Struct TracerEip3155

pub struct TracerEip3155 {
    output: Box<dyn Write>,
    gas_inspector: GasInspector,
    print_summary: bool,
    stack: Vec<Uint<256, 4>>,
    pc: u64,
    opcode: u8,
    gas: u64,
    refunded: i64,
    mem_size: usize,
    skip: bool,
    include_memory: bool,
    memory: Option<String>,
}

EIP-3155 tracer Inspector.

Fields
output: Box<dyn Write>gas_inspector: GasInspectorprint_summary: boolstack: Vec<Uint<256, 4>>pc: u64opcode: u8gas: u64refunded: i64mem_size: usizeskip: boolinclude_memory: boolmemory: Option<String>
Implementations

impl TracerEip3155

pub fn buffered(output: impl Write + 'static) -> TracerEip3155

Creates a new EIP-3155 tracer with the given output writer, by first wrapping it in a BufWriter.

pub fn new_stdout() -> TracerEip3155

Creates a new EIP-3155 tracer with a stdout output.

pub fn new(output: Box<dyn Write>) -> TracerEip3155

Creates a new EIP-3155 tracer with the given output writer.

pub fn set_writer(&mut self, writer: Box<dyn Write>)

Sets the writer to use for the output.

pub fn without_summary(self) -> TracerEip3155

Don’t include a summary at the end of the trace

pub fn with_memory(self) -> TracerEip3155

Include a memory field for each step. This significantly increases processing time and output size.

pub fn clear(&mut self)

Resets the tracer to its initial state of Self::new.

This makes the inspector ready to be used again.