Crate forge::revm::interpreter

revm-interpreter

REVM Interpreter.

Modules

analysis

gas

EVM gas calculation utilities.

instructions

EVM opcode implementations.

primitives

revm-primitives

Macros

push

push_b256

return_error

return_ok

return_revert

Structs

BytecodeLocked

An analyzed bytecode.

CallContext

Context of a runtime call.

CallInputs

Inputs for a call.

CallOutcome

Represents the outcome of a call operation in a virtual machine.

Contract

EVM contract information.

CreateInputs

Inputs for a create call.

CreateOutcome

Represents the outcome of a create operation in an interpreter.

DummyHost

A dummy Host implementation.

Gas

Represents the state of gas during execution.

Interpreter

InterpreterResult

The result of an interpreter operation.

OpCode

An EVM opcode.

SStoreResult

Represents the result of an sstore operation.

SelfDestructResult

Result of a call that resulted in a self destruct.

SharedMemory

A sequential memory shared between calls, which uses a Vec for internal representation. A SharedMemory instance should always be obtained using the new static method to ensure memory safety.

Stack

EVM stack with STACK_LIMIT capacity of words.

Transfer

Transfer from source to target, with given value.

Enums

CallScheme

Call schemes.

CreateScheme

Create scheme.

InstructionResult

InterpreterAction

SuccessOrHalt

Constants

EMPTY_SHARED_MEMORY

Empty shared memory.

MAX_CODE_SIZE

EIP-170: Contract code size limit By default limit is 0x6000 (~25kb)

MAX_INITCODE_SIZE

EIP-3860: Limit and meter initcode

OPCODE_JUMPMAP

Maps each opcode to its name.

STACK_LIMIT

EVM interpreter stack limit.

Traits

Host

EVM context host.

Functions

next_multiple_of_32

Rounds up x to the closest multiple of 32. If x % 32 == 0 then x is returned.

Type Aliases

Instruction

EVM opcode function signature.