Trait forge::revm::interpreter::primitives::alloy_primitives::hex::ToHex

pub trait ToHex {
    // Required methods
    fn encode_hex<T>(&self) -> T
       where T: FromIterator<char>;
    fn encode_hex_upper<T>(&self) -> T
       where T: FromIterator<char>;
}

Encoding values as hex string.

This trait is implemented for all T which implement AsRef<[u8]>. This includes String, str, Vec<u8> and [u8].

Example

use hex::ToHex;

println!("{}", "Hello world!".encode_hex::<String>());

Note: instead of using this trait, you might want to use encode().

Required Methods

fn encode_hex<T>(&self) -> T
where T: FromIterator<char>,

Encode the hex strict representing self into the result. Lower case letters are used (e.g. f9b4ca)

fn encode_hex_upper<T>(&self) -> T
where T: FromIterator<char>,

Encode the hex strict representing self into the result. Upper case letters are used (e.g. F9B4CA)

Object Safety

This trait is not object safe.