forge::revm::interpreter::primitives::alloy_primitives::map

Module hash_set
·

Available on crate feature map only.

A hash set implemented as a HashMap where the value is ().

Structs

Difference

A lazy iterator producing elements in the difference of HashSets.

Drain

A draining iterator over the items of a HashSet.

HashSet

A hash set implemented as a HashMap where the value is ().

Intersection

A lazy iterator producing elements in the intersection of HashSets.

IntoIter

An owning iterator over the items of a HashSet.

Iter

An iterator over the items of a HashSet.

SymmetricDifference

A lazy iterator producing elements in the symmetric difference of HashSets.

Union

A lazy iterator producing elements in the union of HashSets.

ExtractIfExperimental

A draining, filtering iterator over the items of a HashSet.

OccupiedEntryExperimental

A view into an occupied entry in a HashSet. It is part of the Entry enum.

VacantEntryExperimental

A view into a vacant entry in a HashSet. It is part of the Entry enum.

Enums

EntryExperimental

A view into a single entry in a set, which may either be vacant or occupied.