Trait forge::revm::interpreter::primitives::alloy_primitives::ruint::UintTryFrom

pub trait UintTryFrom<T>: Sized { }

⚠️ Workaround for Rust issue #50133. Use TryFrom instead.

We cannot implement TryFrom<Uint> for Uint directly, but we can create a new identical trait and implement it there. We can even give this trait a blanket implementation inheriting all TryFrom<_> implementations.

Object Safety

This trait is not object safe.