Module forge::revm::interpreter::primitives::eof

Structs

Eof

EOF - Ethereum Object Format.

EofBody

EOF Body, contains types, code, container and data sections.

EofHeader

EOF Header containing

TypesSection

Types section that contains stack information for matching code section.

Enums

EofDecodeError

EOF decode errors.

Constants

EOF_MAGIC

EOF Magic in u16 form.

EOF_MAGIC_HASH

Hash of EF00 bytes that is used for EXTCODEHASH when called from legacy bytecode.

Statics

EOF_MAGIC_BYTES

EOF magic number in array form.