forge::revm::interpreter::primitives::kzg
Function parse_kzg_trusted_setup

pub fn parse_kzg_trusted_setup(
    trusted_setup: &str,
) -> Result<(Box<G1Points>, Box<G2Points>), KzgErrors>

Available on crate features c-kzg or kzg-rs only.

Parses the contents of a KZG trusted setup file into a list of G1 and G2 points.

These can then be used to create a KZG settings object with KzgSettings::load_trusted_setup.