Crate forge::revm::precompile

revm-precompile

Implementations of EVM precompiled contracts.

Modules

blake2

bls12_381blst

bn128

hash

identity

kzg_point_evaluationc-kzg

modexp

secp256k1

secp256r1secp256r1

EIP-7212 secp256r1 Precompile

utilities

Structs

Address

An Ethereum address, 20 bytes in length.

Bytes

Wrapper type around [bytes::Bytes] to support “0x” prefixed hex strings.

HashMap

A hash map implemented with quadratic probing and SIMD lookup.

HashSet

A hash set implemented as a HashMap where the value is ().

Log

A log consists of an address, and some log data.

PrecompileOutput

Precompile execution output

PrecompileWithAddress

Precompiles

Enums

Error

Precompile

Precompile and its handlers.

PrecompileError

PrecompileErrors

Precompile errors.

PrecompileSpecId

Traits

StatefulPrecompile

Stateful precompile trait. It is used to create a arc precompile Precompile::Stateful.

StatefulPrecompileMut

Mutable stateful precompile trait. It is used to create a boxed precompile in Precompile::StatefulMut.

Functions

calc_linear_cost_u32

u64_to_address

Const function for making an address by concatenating the bytes from two given numbers.

Type Aliases

B256

32-byte fixed byte-array type.

EnvPrecompileFn

PrecompileResult

A precompile operation result.

StandardPrecompileFn

StatefulPrecompileArc

Arc over stateful precompile.

StatefulPrecompileBox

Box over mutable stateful precompile