forge::revm::precompile::bls12_381

Module g1_add

BLS12-381 G1 add precompile. More details in g1_add

Constants

PRECOMPILE

EIP-2537 BLS12_G1ADD precompile.

Functions

g1_add

G1 addition call expects 256 bytes as an input that is interpreted as byte concatenation of two G1 points (128 bytes each). Output is an encoding of addition operation result - single G1 point (128 bytes). See also: https://eips.ethereum.org/EIPS/eip-2537#abi-for-g1-addition