forge::revm::precompile::bls12_381

Module g2_msm

BLS12-381 G2 msm precompile. More details in g2_msm

Constants

PRECOMPILE

EIP-2537 BLS12_G2MSM precompile.

Functions

g2_msm

Implements EIP-2537 G2MSM precompile. G2 multi-scalar-multiplication call expects 288*k bytes as an input that is interpreted as byte concatenation of k slices each of them being a byte concatenation of encoding of G2 point (256 bytes) and encoding of a scalar value (32 bytes). Output is an encoding of multi-scalar-multiplication operation result - single G2 point (256 bytes). See also: https://eips.ethereum.org/EIPS/eip-2537#abi-for-g2-multiexponentiation