forge::revm::precompile::secp256r1

Function p256_verify_osaka

pub fn p256_verify_osaka(
    input: &[u8],
    gas_limit: u64,
) -> Result<PrecompileOutput, PrecompileError>

secp256r1 precompile logic with Osaka gas cost. It takes the input bytes sent to the precompile and the gas limit. The output represents the result of verifying the secp256r1 signature of the input.

The input is encoded as follows:

signed message hash	r	s	public key x	public key y

32	32	32	32	32