forge::revm::primitives::map

Crate rustc_hash

A speedy, non-cryptographic hashing algorithm used by rustc.

Example

use rustc_hash::FxHashMap;

let mut map: FxHashMap<u32, u32> = FxHashMap::default();
map.insert(22, 44);

Structs

FxBuildHasher

An implementation of BuildHasher that produces FxHashers.

FxHasher

A speedy hash algorithm for use within rustc. The hashmap in liballoc by default uses SipHash which isn’t quite as speedy as we want. In the compiler we’re not really worried about DOS attempts, so we use a fast non-cryptographic hash.

FxRandomState

FxRandomState is an alternative state for HashMap types.

FxSeededState

FxSeededState is an alternative state for HashMap types, allowing to use FxHasher with a set seed.

Type Aliases

FxHashMap

Type alias for a hash map that uses the Fx hashing algorithm.

FxHashMapRand

Type alias for a hashmap using the fx hash algorithm with FxRandomState.

FxHashMapSeed

Type alias for a hashmap using the fx hash algorithm with FxSeededState.

FxHashSet

Type alias for a hash set that uses the Fx hashing algorithm.

FxHashSetRand

Type alias for a hashmap using the fx hash algorithm with FxRandomState.

FxHashSetSeed

Type alias for a hashmap using the fx hash algorithm with FxSeededState.