forge

Module runner

The Forge test runner.

Structs

ContractRunner

A type that executes all tests of a contract

FunctionRunner 🔒

Executes a single test function, returning a TestResult.

InvariantPersistedFailure 🔒

Holds data about a persisted invariant failure.

Constants

LIBRARY_DEPLOYER

When running tests, we deploy all external libraries present in the project. To avoid additional libraries affecting nonces of senders used in tests, we are using separate address to predeploy libraries.

Functions

fuzzer_with_cases 🔒

persisted_call_sequence 🔒

Helper function to load failed call sequence from file. Ignores failure if generated with different test contract than the current one.

test_failure_paths 🔒

Helper functions to return canonicalized test failure paths.