forge

Crate traces

foundry-evm-traces

EVM trace identifying and decoding.

Modules

debug

folded_stack_trace

identifier

Call trace address identifiers.

Structs

CallLog

A log with optional decoded data.

CallTrace

A trace of a call with optional decoded data.

CallTraceArena

An arena of recorded traces.

CallTraceDecoder

The call trace decoder.

CallTraceDecoderBuilder

Build a new CallTraceDecoder.

CallTraceNode

A node in the arena

DebugTraceIdentifier

DecodedCallData

Decoded call data.

DecodedCallLog

Additional decoded data enhancing the CallLog.

DecodedCallTrace

Additional decoded data enhancing the CallTrace.

FourByteInspector

Fourbyte tracing inspector that records all function selectors and their calldata sizes.

GethTraceBuilder

A type for creating geth style traces

ParityTraceBuilder

A type for creating parity style traces

SparsedTraceArena

Trace arena keeping track of ignored trace items.

TraceWriter

Formats call traces to an Write writer.

TracingInspector

An inspector that collects call traces.

TracingInspectorConfig

Gives guidance to the TracingInspector.

Enums

CallKind

A unified representation of a call.

InternalTraceMode

Different kinds of internal functions tracing.

StackSnapshotType

How much of the stack to record. Nothing, just the items pushed, or the full stack

TraceKind

Specifies the kind of trace.

TraceMode

Functions

decode_trace_arena

Decode a collection of call traces.

load_contracts

Given a list of traces and artifacts, it returns a map connecting address to abi

render_trace_arena

Render a collection of call traces to a string.

render_trace_arena_inner

Render a collection of call traces to a string optionally including contract creation bytecodes and in JSON format.

Type Aliases

Traces