forge::traces
Module folded_stack_trace

Structs

EvmFoldedStackTraceBuilder

Wrapper for building a folded stack trace using EVM call trace node.

FoldedStackTraceBuilder

Helps to translate a function enter-exit flow into a folded stack trace.

Functions

build

Builds a folded stack trace from a call trace arena.