forge::traces::identifier

Trait TraceIdentifier

pub trait TraceIdentifier {
    // Required method
    fn identify_addresses(
        &mut self,
        nodes: &[&CallTraceNode],
    ) -> Vec<IdentifiedAddress<'_>>;
}

Trace identifiers figure out what ABIs and labels belong to all the addresses of the trace.

Required Methods

fn identify_addresses( &mut self, nodes: &[&CallTraceNode], ) -> Vec<IdentifiedAddress<'_>>

Attempts to identify an address in one or more call traces.

Dyn Compatibility

This trait is dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety".