Trait forge::NFTPublicUsageTest

pub trait NFTPublicUsageTest: Test {
    fn run<'t, 'life0, 'life1, 'async_trait>(
        &'life0 self, 
        ctx: &'life1 mut NFTPublicUsageContext<'t>
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
    where
        't: 'async_trait,
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;
}

The testing interface which defines a test written from the perspective of the a public user of the NFT network in a “testnet” like environment where there exists a source for minting NFTs and a means of creating new accounts.

Required methods

fn run<'t, 'life0, 'life1, 'async_trait>(
&'life0 self,
ctx: &'life1 mut NFTPublicUsageContext<'t>
) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>> where
't: 'async_trait,
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

Executes the test against the given context.