Trait forge::Swarm

pub trait Swarm {

    fn health_check(&mut self) -> Result<()>;

    fn validators<'a>(
        &'a mut self
    ) -> Box<dyn Iterator<Item = &'a dyn Validator> + 'a>;

    fn validators_mut<'a>(
        &'a mut self
    ) -> Box<dyn Iterator<Item = &'a mut dyn Validator> + 'a>;

    fn validator(&self, id: PeerId) -> Option<&dyn Validator>;

    fn validator_mut(&mut self, id: PeerId) -> Option<&mut dyn Validator>;

    fn full_nodes<'a>(
        &'a mut self
    ) -> Box<dyn Iterator<Item = &'a dyn FullNode> + 'a>;

    fn full_nodes_mut<'a>(
        &'a mut self
    ) -> Box<dyn Iterator<Item = &'a mut dyn FullNode> + 'a>;

    fn full_node(&self, id: PeerId) -> Option<&dyn FullNode>;

    fn full_node_mut(&mut self, id: PeerId) -> Option<&mut dyn FullNode>;

    fn add_validator(&mut self, id: PeerId) -> Result<PeerId>;

    fn remove_validator(&mut self, id: PeerId) -> Result<()>;

    fn add_full_node(&mut self, id: PeerId) -> Result<()>;

    fn remove_full_node(&mut self, id: PeerId) -> Result<()>;

    fn chain_info(&mut self) -> ChainInfo<'_>;

}

Trait used to represent a running network comprised of Validators and FullNodes

Required methods

fn health_check(&mut self) -> Result<()>

Performs a health check on the entire swarm, ensuring all Nodes are Live and that no forks have occurred

fn validators<'a>(
&'a mut self
) -> Box<dyn Iterator<Item = &'a dyn Validator> + 'a>

Returns an Iterator of references to all the Validators in the Swarm

fn validators_mut<'a>(
&'a mut self
) -> Box<dyn Iterator<Item = &'a mut dyn Validator> + 'a>

Returns an Iterator of mutable references to all the Validators in the Swarm

fn validator(&self, id: PeerId) -> Option<&dyn Validator>

Returns a reference to the Validator with the provided PeerId

fn validator_mut(&mut self, id: PeerId) -> Option<&mut dyn Validator>

Returns a mutable reference to the Validator with the provided PeerId

fn full_nodes<'a>(
&'a mut self
) -> Box<dyn Iterator<Item = &'a dyn FullNode> + 'a>

Returns an Iterator of references to all the FullNodes in the Swarm

fn full_nodes_mut<'a>(
&'a mut self
) -> Box<dyn Iterator<Item = &'a mut dyn FullNode> + 'a>

Returns an Iterator of mutable references to all the FullNodes in the Swarm

fn full_node(&self, id: PeerId) -> Option<&dyn FullNode>

Returns a reference to the FullNode with the provided PeerId

fn full_node_mut(&mut self, id: PeerId) -> Option<&mut dyn FullNode>

Returns a mutable reference to the FullNode with the provided PeerId

fn add_validator(&mut self, id: PeerId) -> Result<PeerId>

Adds a Validator to the swarm with the provided PeerId

fn remove_validator(&mut self, id: PeerId) -> Result<()>

Removes the Validator with the provided PeerId

fn add_full_node(&mut self, id: PeerId) -> Result<()>

Adds a FullNode to the swarm with the provided PeerId

fn remove_full_node(&mut self, id: PeerId) -> Result<()>

Removes the FullNode with the provided PeerId

fn chain_info(&mut self) -> ChainInfo<'_>

Construct a ChainInfo from this Swarm

Loading content...