Trait forge::SwarmExt

pub trait SwarmExt: Swarm {
    fn liveness_check<'life0, 'async_trait>(
        &'life0 self, 
        deadline: Instant
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: Sync + 'async_trait,
    { ... }

    fn wait_for_connectivity<'life0, 'async_trait>(
        &'life0 self, 
        deadline: Instant
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: Sync + 'async_trait,
    { ... }

    fn fork_check(&self) -> Result<()> { ... }

    fn wait_for_all_nodes_to_catchup_to_version<'life0, 'async_trait>(
        &'life0 self, 
        version: u64, 
        deadline: Instant
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: Sync + 'async_trait,
    { ... }

    fn wait_for_all_nodes_to_catchup<'life0, 'async_trait>(
        &'life0 self, 
        deadline: Instant
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: Sync + 'async_trait,
    { ... }
}

Provided methods

fn liveness_check<'life0, 'async_trait>(
&'life0 self,
deadline: Instant
) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: Sync + 'async_trait,

fn wait_for_connectivity<'life0, 'async_trait>(
&'life0 self,
deadline: Instant
) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: Sync + 'async_trait,

Waits for the swarm to achieve connectivity

fn fork_check(&self) -> Result<()>

Perform a safety check, ensuring that no forks have occurred in the network.

fn wait_for_all_nodes_to_catchup_to_version<'life0, 'async_trait>(
&'life0 self,
version: u64,
deadline: Instant
) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: Sync + 'async_trait,

Waits for all nodes to have caught up to the specified verison.

fn wait_for_all_nodes_to_catchup<'life0, 'async_trait>(
&'life0 self,
deadline: Instant
) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: Sync + 'async_trait,

Wait for all nodes in the network to be caught up. This is done by first querying each node for its current version, selects the max version, then waits for all nodes to catch up to that version. Once done, we can guarantee that all transactions committed before invocation of this function are available at all the nodes in the swarm