Crate forge

Forge is a framework for writing and running end-to-end tests in Diem

Modules

atomic_histogram

cluster

instance

Structs

AdminContext

Author

ChainInfo

CommitInfo

CoreContext

EmitJob

EmitJobRequest

EmitThreadParams

Forge

ForgeConfig

GitCommitInfo

GitHub

K8sFactory

K8sNode

K8sSwarm

KubeService

LocalFactory

LocalNode

LocalSwarm

LocalSwarmBuilder

LocalVersion

NFTPublicInfo

NFTPublicUsageContext

NetworkContext

Options

PublicInfo

PublicUsageContext

ReportedMetric

SlackClient

TestReport

TxnEmitter

TxnStats

TxnStatsRate

Version

A wrapper around a usize in order to represent an opaque version of a Node.

Enums

Coffer

Format

GenesisConfig

HealthCheckError

InitialVersion

ShouldFail

Whether a test is expected to fail or not

SwarmDirectory

Traits

AdminTest

The testing interface which defines a test written from the perspective of the Admin of the network. This means that the test will have access to the Root account but do not control any of the validators or full nodes running on the network.

Factory

Trait used to represent a interface for constructing a launching new networks

FullNode

Trait used to represent a running FullNode

Fund

NFTPublicUsageTest

The testing interface which defines a test written from the perspective of the a public user of the NFT network in a “testnet” like environment where there exists a source for minting NFTs and a means of creating new accounts.

NetworkTest

The testing interface which defines a test written with full control over an existing network. Tests written against this interface will have access to both the Root account as well as the nodes which comprise the network.

Node

Trait used to represent a running Validator or FullNode

NodeExt

PublicUsageTest

The testing interface which defines a test written from the perspective of the a public user of the network in a “testnet” like environment where there exists a funding source and a means of creating new accounts.

Swarm

Trait used to represent a running network comprised of Validators and FullNodes

SwarmExt

Test

Represents a Test in Forge

Validator

Trait used to represent a running Validator

Functions

clean_k8s_cluster

create_k8s_client

create_parent_vasp_request

execute_and_wait_transactions

forge_main

gen_transfer_txn_request

nodes_healthcheck

query_sequence_numbers

scale_sts_replica

set_eks_nodegroup_size

set_validator_image_tag

uninstall_from_k8s_cluster

Type Definitions

Result

Result<T, Error>