Crate forge

Forge is a framework for writing and running end-to-end tests in Diem

Modules

atomic_histogram

Structs

AdminContext

ChainInfo

CoreContext

EmitJob

EmitJobRequest

EmitThreadParams

Forge

ForgeConfig

K8sFactory

K8sNode

K8sSwarm

KubeService

LocalFactory

LocalNode

LocalSwarm

LocalSwarmBuilder

NetworkContext

Options

PublicInfo

PublicUsageContext

ReportedMetric

TestReport

TxnEmitter

TxnStats

TxnStatsRate

Enums

Coffer

Format

HealthCheckError

ShouldFail
Whether a test is expected to fail or not

Traits

AdminTest
The testing interface which defines a test written from the perspective of the Admin of the network. This means that the test will have access to the Root account but do not control any of the validators or full nodes running on the network.

Factory
Trait used to represent a interface for constructing a launching new networks

FullNode
Trait used to represent a running FullNode

Fund

NetworkTest
The testing interface which defines a test written with full control over an existing network. Tests written against this interface will have access to both the Root account as well as the nodes which comprise the network.

Node
Trait used to represent a running Validator or FullNode

NodeExt

PublicUsageTest
The testing interface which defines a test written from the perspective of the a public user of the network in a “testnet” like environment where there exists a funding source and a means of creating new accounts.

Swarm
Trait used to represent a running network comprised of Validators and FullNodes

Test
Represents a Test in Forge

Validator
Trait used to represent a running Validator

Functions

create_parent_vasp_request

execute_and_wait_transactions

forge_main

gen_transfer_txn_request

list_services

query_sequence_numbers

Type Definitions

Result
Result<T, Error>