Crate forge

Modules

backend

Foundry’s main executor backend abstraction and implementation.

constants

coverage

Coverage reports.

debug

decode

Various utilities to decode test results.

executors

EVM executor abstractions, which can execute calls.

fork

fuzz

foundry-evm-fuzz

gas_report

Gas reports.

inspectors

EVM inspectors.

multi_runner 🔒

Forge test runner for multiple contracts.

opts

result

Test outcomes.

revm

Revm is a Rust EVM implementation.

runner 🔒

The Forge test runner.

traces

foundry-evm-traces

utils

Macros

call_inspectors

Helper macro to call the same method on multiple inspectors without resorting to dynamic dispatch.

Structs

ContractRunner

A type that executes all tests of a contract

MultiContractRunner

A multi contract runner receives a set of contracts deployed in an EVM instance and proceeds to run all test functions in these contracts.

MultiContractRunnerBuilder

Builder used for instantiating the multi-contract runner

TestOptions

Metadata on how to run fuzz/invariant tests

TestOptionsBuilder

Builder utility to create a TestOptions instance.

Traits

TestFilter

Test filter.