Crate forge

Forge is a fast and flexible Ethereum testing framework.

Re-exports

pub use multi_runner::MultiContractRunner;

pub use multi_runner::MultiContractRunnerBuilder;

Modules

args

backend

Foundry’s main executor backend abstraction and implementation.

cmd

forge subcommands.

constants

coverage

Coverage reports.

decode

Various utilities to decode test results.

executors

EVM executor abstractions, which can execute calls.

fork

fuzz

foundry-evm-fuzz

gas_report

Gas reports.

inspectors

EVM inspectors.

lockfile 🔒

foundry.lock handler type.

multi_runner

Forge test runner for multiple contracts.

opts

progress 🔒

result

Test outcomes.

revm

Revm is a Rust EVM implementation.

runner 🔒

The Forge test runner.

traces

foundry-evm-traces

utils

Macros

call_inspectors

Helper macro to call the same method on multiple inspectors without resorting to dynamic dispatch.

Structs

ContractRunner

A type that executes all tests of a contract

Env

Helper container type for EvmEnv and TxEnv.

EnvMut

Helper struct with mutable references to the block and cfg environments.

EvmEnv

Container type that holds both the configuration and block environment for EVM execution.

Lockfile

A lockfile handler that keeps track of the dependencies and their current state.

Enums

DepIdentifier

Identifies whether a dependency (submodule) is referenced by a branch, tag or rev (commit hash).

Constants

FOUNDRY_LOCK

Traits

InspectorExt

An extension trait that allows us to add additional hooks to Inspector for later use in handlers.

TestFilter

Test filter.

Type Aliases

DepMap

A type alias for a HashMap of dependencies keyed by relative path to the submodule dir.