forge

Module coverage

Coverage reports.

Modules

analysis

anchors

Structs

BytecodeReporter

ContractId

A unique identifier for a contract

CoverageItem

CoverageReport

A coverage report.

CoverageSummary

Coverage summary for a source file.

CoverageSummaryReporter

A simple summary reporter that prints the coverage results in a table.

DebugReporter

A super verbose reporter for debugging coverage while it is still unstable.

HitMap

Hit data for an address.

HitMaps

A collection of HitMaps.

ItemAnchor

An item anchor describes what instruction marks a CoverageItem as covered.

LcovReporter

Writes the coverage report in LCOV’s tracefile format.

LineCoverageCollector

Inspector implementation for collecting coverage information.

LineNumberCache 🔒

Cache line number offsets for source files

SourceLocation

A source location.

Enums

CoverageItemKind

Traits

CoverageReporter

A coverage reporter.

Functions

format_cell 🔒