forge

Crate fuzz

foundry-evm-fuzz

EVM fuzzing implementation using [proptest].

Modules

invariant

strategies

Structs

BaseCounterExample

BasicTxDetails

Details of a transaction generated by fuzz strategy for fuzzing a target.

CallDetails

Call details of a transaction generated to fuzz.

FuzzCase

Data of a single fuzz test case

FuzzConfig

Configuration for how a proptest test should be run.

FuzzFixtures

Fixtures to be used for fuzz tests.

FuzzTestResult

The outcome of a fuzz test

FuzzedCases

Container type for all successful test cases

Fuzzer

An inspector that can fuzz and collect data for that effect.

Reason

The reason for why something, such as a generated value, was rejected.

Enums

CounterExample

FuzzError

Possible errors when running fuzz tests

Functions

fixture_name

Extracts fixture name from a function name. For example: fixtures defined in fixture_Owner function will be applied for owner parameter.