forge

Module result

Test outcomes.

Macros

extend 🔒

Structs

SuiteResult

A set of test results for a single test suite, which is all the tests in a single contract.

SuiteTestResult

The result of a single test in a test suite.

TestOutcome

The aggregated result of a test run.

TestResult

The result of an executed test.

TestSetup

The result of a test setup.

Enums

TestKind

Various types of tests

TestKindReport

Data report by a test.

TestStatus

The status of a test.