forge

Crate revm

Revm is a Rust EVM implementation.

Modules

bytecode

Crate that contains bytecode types and opcode constants.

context

EVM execution context.

context_interface

EVM execution context interface.

database

Database implementations.

database_interface

Database interface.

handler

EVM execution handling.

inspector

Inspector is a crate that provides a set of traits that allow inspecting the EVM execution.

interpreter

revm-interpreter

precompile

revm-precompile

primitives

revm-primitives

state

Account and storage state.

Structs

Context

EVM context contains data that EVM needs for execution.

Journal

A journal of state changes internal to the EVM

Enums

JournalEntry

Journal entries that are used to track changes to the state and are used to revert it.

Traits

Database

EVM database interface.

DatabaseCommit

EVM database commit interface.

DatabaseRef

EVM database interface.

ExecuteCommitEvm

Extension of the ExecuteEvm trait that adds a method that commits the state after execution.

ExecuteEvm

Execute EVM transactions. Main trait for transaction execution.

InspectCommitEvm

InspectCommitEvm is a API that allows inspecting similar to InspectEvm but it has functions that commit the state diff to the database.

InspectEvm

InspectEvm is a API that allows inspecting the EVM.

InspectSystemCallEvm

InspectSystemCallEvm is an API that allows inspecting system calls in the EVM.

Inspector

EVM hooks into execution.

MainBuilder

Trait for building mainnet EVM instances from contexts.

MainContext

Trait used to initialize Context with default mainnet types.

SystemCallCommitEvm

Extension of the SystemCallEvm trait that adds a method that commits the state after execution.

SystemCallEvm

API for executing the system calls. System calls dont deduct the caller or reward the beneficiary. They are used before and after block execution to insert or obtain blockchain state.

Functions

install_crypto

Install a custom crypto provider globally.

Type Aliases

MainnetEvm

Type alias for a mainnet EVM instance with standard Ethereum components.