forge::revm

Crate bytecode

Crate that contains bytecode types and opcode constants.

Legacy bytecode will always contain a jump table.

While EIP-7702 bytecode must contains a Address.

Modules

bitvec

bytecode

Module that contains the bytecode enum with all variants supported by Ethereum mainnet.

eip7702

EIP-7702 bytecode.

legacy

Legacy bytecode.

opcode

EVM opcode definitions and utilities. It contains opcode information and utilities to work with opcodes.

utils

Various utilities for the bytecode

Structs

BytecodeIterator

Iterator over opcodes in a bytecode, skipping immediates.

JumpTable

A table of valid jump destinations.

LegacyAnalyzedBytecode

Legacy analyzed bytecode represents the original bytecode format used in Ethereum.

LegacyRawBytecode

Used only as intermediate representation for legacy bytecode.

OpCode

An EVM opcode

Enums

Bytecode

Main bytecode structure with all variants.

BytecodeDecodeError

Bytecode decode errors