Crate forge::revm::interpreter

revm-interpreter

REVM Interpreter.

Modules

gas

EVM gas calculation utilities.

instructions

EVM opcode implementations.

interpreter

primitives

revm-primitives

Macros

as_u64_saturated

Converts a U256 value to a u64, saturating to MAX if the value is too large.

as_usize_or_fail

Converts a U256 value to a usize, failing the instruction if the value is too large.

as_usize_or_fail_ret

Converts a U256 value to a usize and returns ret, failing the instruction if the value is too large.

as_usize_saturated

Converts a U256 value to a usize, saturating to MAX if the value is too large.

check

Fails the instruction if the min is not enabled in SPEC.

check_staticcall

Fails the instruction if the current call is static.

gas

Records a gas cost and fails the instruction if it would exceed the available gas.

gas_or_fail

Same as gas!, but with gas as an option.

pop

Pops U256 values from the stack. Fails the instruction if the stack is too small.

pop_address

Pops Address values from the stack. Fails the instruction if the stack is too small.

pop_ret

Pops U256 values from the stack, and returns ret. Fails the instruction if the stack is too small.

pop_top

Pops U256 values from the stack, and returns a reference to the top of the stack. Fails the instruction if the stack is too small.

push

Pushes a B256 value onto the stack. Fails the instruction if the stack is full.

push_b256

Pushes B256 values onto the stack. Fails the instruction if the stack is full.

refund

Records a gas refund.

resize_memory

Resizes the interpreter memory if necessary. Fails the instruction if the memory or gas limit is exceeded.

return_error

return_ok

return_revert

Structs

BytecodeLocked

An analyzed bytecode.

CallContext

Context of a runtime call.

CallInputs

Inputs for a call.

CallOutcome

Represents the outcome of a call operation in a virtual machine.

Contract

EVM contract information.

CreateInputs

Inputs for a create call.

CreateOutcome

Represents the outcome of a create operation in an interpreter.

DummyHost

A dummy Host implementation.

Gas

Represents the state of gas during execution.

Interpreter

EVM bytecode interpreter.

InterpreterResult

The result of an interpreter operation.

OpCode

An EVM opcode.

SStoreResult

Represents the result of an sstore operation.

SelfDestructResult

Result of a call that resulted in a self destruct.

SharedMemory

A sequential memory shared between calls, which uses a Vec for internal representation. A SharedMemory instance should always be obtained using the new static method to ensure memory safety.

Stack

EVM stack with STACK_LIMIT capacity of words.

Transfer

Transfer from source to target, with given value.

Enums

CallScheme

Call schemes.

CreateScheme

Create scheme.

InstructionResult

InterpreterAction

SuccessOrHalt

Constants

EMPTY_SHARED_MEMORY

Empty shared memory.

MAX_CODE_SIZE

EIP-170: Contract code size limit By default limit is 0x6000 (~25kb)

MAX_INITCODE_SIZE

EIP-3860: Limit and meter initcode

OPCODE_JUMPMAP

Maps each opcode to its name.

STACK_LIMIT

EVM interpreter stack limit.

Traits

Host

EVM context host.

Functions

next_multiple_of_32

Rounds up x to the closest multiple of 32. If x % 32 == 0 then x is returned. Note, if x is greater than usize::MAX - 31 this will return usize::MAX which isn’t a multiple of 32.

Type Aliases

Instruction

EVM opcode function signature.