forge::revm::interpreter::primitives::bitvec
Module prelude

Symbol Export

This module collects the general public API into a single place for bulk import, as use bitvec::prelude::*;, without polluting the root namespace of the crate.

This provides all the data structure types and macros, as well as the two traits needed to operate them as type parameters, by name. It also imports extension traits without naming them, so that their methods are available but their trait names are not.

Macros

BitArr

Bit-Array Type Definition

bitarr

Bit-Array Value Constructor

bitboxalloc

Boxed Bit-Slice Constructor

bits

Bit-Slice Region Constructor

bitvecalloc

Bit-Vector Constructor

Structs

BitArray

Bit-Precision Array Immediate

BitBox

Fixed-Size, Heap-Allocated, Bit Slice

BitPtr

Single-Bit Pointer

BitPtrRange

Bit-Pointer Range

BitRef

Proxy Bit-Reference

BitSlice

Bit-Addressable Memory

BitVec

Bit-Precision Dynamic Array

LocalBits

Least-Significant-First Bit Traversal

Lsb0

Least-Significant-First Bit Traversal

Msb0

Most-Significant-First Bit Traversal

Traits

AsBits

Immutable Bit View

AsMutBits

Mutable Bit View

BitOrder

In-Element Bit Ordering

BitStore

Bit Storage

_

C-Style Bit-Field Access

_

Bit View

_

Helper trait for scalars and arrays, but not slices.