Macro forge::revm::interpreter::primitives::bitvec::prelude::bitvec

macro_rules! bitvec {
    ($store:ty, $order:ty; $val:expr; $len:expr) => { ... };
    (Cell<$store:ident>, $order:ident $($rest:tt)*) => { ... };
    ($val:expr; $len:expr) => { ... };
    ($($arg:tt)*) => { ... };
}

Available on crate feature alloc only.

Bit-Vector Constructor

This macro creates encoded BitSlice buffers at compile-time, and at run-time copies them directly into a new heap allocation.

It forwards all of its arguments to bits!, and calls BitVec::from_bitslice on the produced &BitSlice expression. While you can use the bits! modifiers, there is no point, as the produced bit-slice is lost before the macro exits.