forge::revm::interpreter::primitives::bitvec::ptr

Function drop_in_place

pub unsafe fn drop_in_place<T, O>(_: BitPtr<Mut, T, O>)
where
    T: BitStore,
    O: BitOrder,

👎Deprecated: this has no effect, and should not be called

Remote Destructor

BitPtr only points to indestructible types. This has no effect, and is only present for symbol compatibility. You should not have been calling it on your integers or bools anyway!

Original

ptr::drop_in_place