Crate forge::traces

foundry-evm-traces

EVM trace identifying and decoding.

Modules

identifier

Call trace address identifiers.

Structs

CallTrace

A trace of a call with optional decoded data.

CallTraceArena

An arena of recorded traces.

CallTraceDecoder

The call trace decoder.

CallTraceDecoderBuilder

Build a new CallTraceDecoder.

CallTraceNode

A node in the arena

DecodedCallData

DecodedCallTrace

FourByteInspector

Fourbyte tracing inspector that records all function selectors and their calldata sizes.

GethTraceBuilder

A type for creating geth style traces

ParityTraceBuilder

A type for creating parity style traces

TracingInspector

An inspector that collects call traces.

TracingInspectorConfig

Gives guidance to the TracingInspector.

Enums

CallKind

A unified representation of a call.

DecodedCallLog

StackSnapshotType

How much of the stack to record. Nothing, just the items pushed, or the full stack

TraceKind

Specifies the kind of trace.

Functions

load_contracts

Given a list of traces and artifacts, it returns a map connecting address to abi

render_trace

Render a call trace.

render_trace_arena

Render a collection of call traces.

Type Aliases

Traces