Trait forge::Node

pub trait Node {
    fn peer_id(&self) -> PeerId;

    fn name(&self) -> &str;

    fn json_rpc_endpoint(&self) -> Url;

    fn debug_endpoint(&self) -> Url;

    fn config(&self) -> &NodeConfig;

    fn start(&mut self) -> Result<()>;

    fn stop(&mut self) -> Result<()>;

    fn clear_storage(&mut self) -> Result<()>;

    fn health_check(&mut self) -> Result<(), HealthCheckError>;

    fn restart(&mut self) -> Result<()> { ... }
}

Trait used to represent a running Validator or FullNode

Required methods

fn peer_id(&self) -> PeerId

Return the PeerId of this Node

fn name(&self) -> &str

Return the human readable name of this Node

fn json_rpc_endpoint(&self) -> Url

Return the URL for the JSON-RPC endpoint of this Node

fn debug_endpoint(&self) -> Url

Return the URL for the debug-interface for this Node

fn config(&self) -> &NodeConfig

Return a reference to the Config this Node is using

fn start(&mut self) -> Result<()>

Start this Node. This should be a noop if the Node is already running.

fn stop(&mut self) -> Result<()>

Stop this Node. This should be a noop if the Node isn’t running.

fn clear_storage(&mut self) -> Result<()>

Clears this Node’s Storage

fn health_check(&mut self) -> Result<(), HealthCheckError>

Performs a Health Check on the Node

Loading content...
Provided methods

fn restart(&mut self) -> Result<()>

Restarts this Node by calling Node::Stop followed by Node::Start

Loading content...