Trait forge::Swarm

pub trait Swarm {
    fn health_check(&mut self) -> Result<()>;

    fn validators<'a>(
        &'a mut self
    ) -> Box<dyn Iterator<Item = &'a dyn Validator> + 'a>;

    fn validators_mut<'a>(
        &'a mut self
    ) -> Box<dyn Iterator<Item = &'a mut dyn Validator> + 'a>;

    fn validator(&self, id: NodeId) -> &dyn Validator;

    fn validator_mut(&mut self, id: NodeId) -> &mut dyn Validator;

    fn full_nodes<'a>(
        &'a mut self
    ) -> Box<dyn Iterator<Item = &'a dyn FullNode> + 'a>;

    fn full_nodes_mut<'a>(
        &'a mut self
    ) -> Box<dyn Iterator<Item = &'a mut dyn FullNode> + 'a>;

    fn full_node(&self, id: NodeId) -> &dyn FullNode;

    fn full_node_mut(&mut self, id: NodeId) -> &mut dyn FullNode;

    fn add_validator(&mut self, id: NodeId) -> Result<NodeId>;

    fn remove_validator(&mut self, id: NodeId) -> Result<()>;

    fn add_full_node(&mut self, id: NodeId) -> Result<()>;

    fn remove_full_node(&mut self, id: NodeId) -> Result<()>;

    fn admin_info(&mut self) -> AdminInfo<'_>;

    fn public_info(&mut self) -> PublicInfo<'_>;

    fn chain_info(&mut self) -> ChainInfo<'_>;
}

Trait used to represent a running network comprised of Validators and FullNodes

Required methods

fn health_check(&mut self) -> Result<()>

Performs a health check on the entire swarm, ensuring all Nodes are Live and that no forks have occurred

fn validators<'a>(
&'a mut self
) -> Box<dyn Iterator<Item = &'a dyn Validator> + 'a>

Returns an Iterator of references to all the Validators in the Swarm

fn validators_mut<'a>(
&'a mut self
) -> Box<dyn Iterator<Item = &'a mut dyn Validator> + 'a>

Returns an Iterator of mutable references to all the Validators in the Swarm

fn validator(&self, id: NodeId) -> &dyn Validator

Returns a reference to the Validator with the provided NodeId

fn validator_mut(&mut self, id: NodeId) -> &mut dyn Validator

Returns a mutable reference to the Validator with the provided NodeId

fn full_nodes<'a>(
&'a mut self
) -> Box<dyn Iterator<Item = &'a dyn FullNode> + 'a>

Returns an Iterator of references to all the FullNodes in the Swarm

fn full_nodes_mut<'a>(
&'a mut self
) -> Box<dyn Iterator<Item = &'a mut dyn FullNode> + 'a>

Returns an Iterator of mutable references to all the FullNodes in the Swarm

fn full_node(&self, id: NodeId) -> &dyn FullNode

Returns a reference to the FullNode with the provided NodeId

fn full_node_mut(&mut self, id: NodeId) -> &mut dyn FullNode

Returns a mutable reference to the FullNode with the provided NodeId

fn add_validator(&mut self, id: NodeId) -> Result<NodeId>

Adds a Validator to the swarm with the provided NodeId

fn remove_validator(&mut self, id: NodeId) -> Result<()>

Removes the Validator with the provided NodeId

fn add_full_node(&mut self, id: NodeId) -> Result<()>

Adds a FullNode to the swarm with the provided NodeId

fn remove_full_node(&mut self, id: NodeId) -> Result<()>

Removes the FullNode with the provided NodeId

fn admin_info(&mut self) -> AdminInfo<'_>

Construct an AdminInfo from this Swarm

fn public_info(&mut self) -> PublicInfo<'_>

Construct a PublicInfo from this Swarm

fn chain_info(&mut self) -> ChainInfo<'_>

Construct a NetworkInfo from this Swarm

Loading content...